### 2018-06-28 #233 @kamapu
taxlist, a package for structuring taxonomic lists and related information
The `taxlist` package structures taxonomic information into **S4** objects and implements methods for the manipulation of contained information.
Such objects may or may not contain information on synonymy, taxonomic ranks, parent-child relations, taxon views (references used to establish relation between taxon usage names and taxon concepts), and taxon (functional) traits.

```
Package: taxlist
Version: 0.1.5
Title: Handling Taxonomic Lists
	person("Miguel", "Alvarez", email="kamapu78@gmail.com", role=c("aut", "cre"))
    R(>= 3.0.0),
    utils
    foreign,
    grDevices,
    taxize,
    stringdist,
    vegdata
	ape,
    devtools,
    stringi,
    taxa,
Description: Handling taxonomic lists through objects of class 'taxlist'.
    This package provides functions to import species lists from 'Turboveg'
    (<https://www.synbiosys.alterra.nl/turboveg>) and the possibility to create
    backups from resulting R-objects.
    Also quick displays are implemented as summary-methods.
URL:
	https://cran.r-project.org/package=taxlist,
	https://github.com/kamapu/taxlist
Collate:
    'NULLing.R''auxiliary_functions.R''deprecated-functions.R''dissect_name.R'
    'taxlist-class.R''clean.R''as.list.R''taxon_views.R''add_view.R'
    'taxon_names.R''taxon_relations.R''taxon_traits.R'
    'levels.R''add_concept.R''update_concept.R''add_synonym.R'
    'accepted_name.R''synonyms.R''basionym.R''update_name.R''delete_name.R'
    'replace_view.R''get_children.R'
    'change_concept.R''Extract.R''subset.R'
    'merge_taxa.R''backup_object.R''load_last.R''summary.R'
    'df2taxlist.R''tv2taxlist.R''tnrs.R''tax2traits.R''match_names.R''print_name.R'
    'StartMessage.R'
```

https://github.com/kamapu/taxlist

- Please indicate which category or categories from our [package fit policies](https://github.com/ropensci/onboarding/blob/master/policies.md#package-fit) this package falls under ***and why?** (e.g., data retrieval, reproducibility. If you are unsure, we suggest you make a pre-submission inquiry.):

Reproducibility, because this package makes taxonomic information available in a *quasi-standard* format and tests inconsistencies on the content of taxonomic lists.

In general to taxonomists and biodiversity scientists, in particular to vegetation ecologists (`taxlist` objects are implemented in the package [vegtable](https://CRAN.R-project.org/package=vegtable)).

While its functionality may overlap the package [taxa](https://cran.r-project.org/package=taxa), the package `taxlist` attempts to be flexible in the degree of completeness of data (incompleteness is very frequent in vegetation-plot databases), it is meant to be integrated in objects containing diversity information (as in the mentioned package `vegtable`) and to import data from local storage (spreadsheets, Turboveg data sets and even PostgreSQL tables by using [vegtable2](https://github.com/kamapu/vegtable2)).

@arendsee
@zachary-foster
@sckott
