### 2018-06-28 3e9d728
Some minor adjustments in documentation
