### 2018-06-29 ab6f958
option style="knitr" in 'print_name'
