### 2018-07-02 3f8d73f
Message in 'backup_object' after saving
