+++ NEWS.md
@@ -2,7 +2,7 @@ taxlist 0.1.6
 =============
 
 ### Improvements
-* A new option `style="knitr"` for function `print_name()`.
+* A new option `style="knitr"` for function `print_name()` (See [this issue](https://stackoverflow.com/questions/51092103/formatted-scientific-names-from-r-to-latex-using-sweave-or-knitr)) at **Stack Overflow**.
 * In function `backup_object`, the message will be done after successful saving and not before.
 
 taxlist 0.1.5
