### 2018-07-03 c7250c0
Small change in documentation of 'as.phylo'
Dollar symbol in example have to be escaped
