### 2018-07-04 f6e5933
Bug in function 'add_synonym' (see NEWS)
