### 2018-07-05 e9fd439
Small bug in function 'load_last'
