### 2018-07-07 97122cc
New function 'clean_strings'
