+++ TODO.md
@@ -1,4 +1,4 @@
 
 * Improve documentation in `print_name`
 * Solve problems with `as.phylo` (communication with developer of package `ape`)
-* Check for missing
+* Check for function `tnrs` and skip it from vignette
