+++ DESCRIPTION
@@ -35,7 +35,8 @@ URL:
 	https://github.com/kamapu/taxlist
 BugReports: https://github.com/kamapu/taxlist/issues
 Collate:
-    'NULLing.R''auxiliary_functions.R''deprecated-functions.R''dissect_name.R'
+    'NULLing.R''auxiliary_functions.R''deprecated-functions.R'
+    'dissect_name.R''clean_strings.R'
     'taxlist-class.R''clean.R''as.list.R''taxon_views.R''add_view.R'
     'taxon_names.R''taxon_relations.R''taxon_traits.R'
     'levels.R''add_concept.R''update_concept.R''add_synonym.R'
