### 2018-07-07 47b6900
function 'tnrs' excluded from vignette
