+++ TODO.md
@@ -1,4 +1,5 @@
-
 * Improve documentation in `print_name`
 * Solve problems with `as.phylo` (communication with developer of package `ape`)
-* Check for function `tnrs` and skip it from vignette
+* Check for function `tnrs`
+* Make function `check_names` more efficient for big species lists
+* In `match_names` an argument to compare only among accepted names
