### 2018-07-07 92b3a52
file.path not working properly inside of functions
