### 2018-07-09 0c1a054
function 'levels' has to be changed to 'base::levels'
