### 2018-07-10 #233 @maelle
:wave: again @kamapu, it's a bit of a tricky situation and we're going to need more information about your package before making a decision.

The general question is: **how would an R user with a taxonomy problem at hand know how to choose one of the two packages?**

Your answer could be a table [as in this pre-submission inquiry](https://github.com/ropensci/onboarding/issues/199#issuecomment-375358362) or a decision tree. I know it's a bit of work that might not even warrant onboarding, but it'd be something useful to have in the docs of your package (and of `taxa` probably) since the existence of both packages even not under the rOpenSci's umbrella mean users might have to make a decision.

Thanks for your patience!
