### 2018-09-03 #233 @kamapu
Dear @maelle, there are not really updates on this issue. At the moment I am for a longer period in the field (Kenya). Though, it will be difficult for me to really document differences between `taxlist` and `taxa`. I only can tell in favour of `taxlist` that this package is implemented in `vegtable`, which is meant as a container for vegetation-plot databases in **R**. In fact, `taxlist` development runs in parallel to [`vegtable`](https://github.com/kamapu/vegtable) but they are kept in separated packages due to the complexity of taxonomic structures and the potential integration of taxonomic lists in other object classes.

Now my question is: Can I submit two packages as a bulk in **ROpenSci** or should I leave `taxlist` and try to submit `vegtable` alone?

Best regards!
