### 2018-10-08 #233 @kamapu
Dear @maelle:

I went through most of the functions in the manual of [`taxa`](https://github.com/ropensci/taxa) to have a better impression about differences between this package and [`taxlist`](https://github.com/kamapu/taxlist). I have to recognize, it has been hard for me to understand the way how `taxa` deals with taxonomic information, since its programming style is a higher level than my primitive programming skills. Thus, the following comments should be interpreted as impressions rather than truth.

Basically both mentioned packages fulfil similar tasks but use different approaches, `taxa` uses a *function-oriented* approach (`R6`), while `taxlist` uses a *data-oriented* one (`S4`). Both packages attempt to provide an object class containing taxonomic information, but `taxlist` is strongly focusing on potential applications in vegetation-plot databases. `taxlist` allows diverse degrees of completeness, starting from just listing of taxonomic entities up to data sets including taxonomic ranks, synonyms,
ecological traits (or further taxon attributes), and taxon views.

## Contributions of taxlist

Further aspects are probably not implemented in `taxa` an can be considered as the contribution of `taxlist` to **ROpenSci**:

- While `taxa` is importing information from several on-line databases,   `taxlist` offers alternatives of import data from `local` data sets. For   instance data loaded in data frames can be converted by using the function `df2taxlist()`. There is also a function called `tv2taxlist()` for importing species lists included in [**Turboveg**](https://www.synbiosys.alterra.nl/turboveg)
- The consistency of information included in taxonomic lists is "cross-checked" by *validity* in `S4` objects. In the case of `taxlist`, potential issues such as duplicated combinations, parent-child relationships that are not consistent to taxonomic ranks, etc. will generate error messages when creating new objects. It is not clear for me how those issues are handled in `taxa`
- I don't see any mentions to *synonymy* in `taxa`, which is a crucial feature in `taxlist`, assuming that all databases including data from different sources have to implement a way of linking taxon usage names (either accepted names or synonyms) with taxon concepts. There is no mention to *taxon views* in `taxa`, either
- A function `backup_object()` for inserting time stamps to *backups* produced by `save()`
- Some string process and comparisons through the functions `dissect_name()` (split names into single terms) and `match_names()` for comparing character values with names included in `taxlist` objects
- Interaction with `knitr` and `rmarkdown` for inserting taxon names in respective documents using the function `print_name()`

## Features in Development

Additional features in development, which may not be yet included in `taxa` are:

- An object class for multiple taxon views (`taxlist` enable single taxon views per taxon concept).
- A plotting function displaying relations between taxon as dendrograms
- A function exporting the content of `taxlist` objects into the [**Veg-X**](http://bien.nceas.ucsb.edu/bien/tools/vegx) standard

## Final Remarks

I hope the previous comments properly support `taxlist` as a package rather complementary to `taxa` than redundant and make it eligible for **ROpenSci**. Disregarding the final decision, a way to make data sets exchangeable among those two applications should be strongly recommended.
