### 2018-10-08 #233 @zachary-foster
Thanks for looking through the docs/code @kamapu! Here are some thoughts:

> While taxa is importing information from several on-line databases, taxlist offers alternatives of import data from local data sets.

`taxa` can read local datasets using `parse_tax_data`, but it does not have any parsers for specific local formats, since we wanted to keep everything generic.

> The consistency of information included in taxonomic lists is "cross-checked" by validity in S4 objects.

Yep, `taxa` uses R6 classes which are not strong typed, so users can break the objects if they set fields manually instead of using our dplyr-like manipulation functions.

> I don't see any mentions to synonymy in taxa, which is a crucial feature in taxlist

There is no such feature explicitly and we dont have plans for one at the moment

>  A plotting function displaying relations between taxon as dendrograms

`taxa` uses `print_tree` for this

> A function exporting the content of taxlist objects into the Veg-X standard

I have no plans to support import or export of specific file formats in `taxa`, since it is primarily a data manipulation standard.
