### 2018-10-16 bb5498d
function 'accepted_name' retrieving NA traits
