+++ NEWS.md
@@ -6,12 +6,13 @@ taxlist 0.1.6
 * New methods `as.phylo` and `as.multiPhylo` for plotting parent-child relations in dendrograms
 
 ### Improvements
-* A new option `style="knitr"` for function `print_name` (See [this issue](https://stackoverflow.com/questions/51092103/formatted-scientific-names-from-r-to-latex-using-sweave-or-knitr)) at **Stack Overflow**.
-* In function `backup_object`, the message will be done after successful saving and not before.
+* A new option `style="knitr"` for function `print_name()` (See [this issue](https://stackoverflow.com/questions/51092103/formatted-scientific-names-from-r-to-latex-using-sweave-or-knitr)) at **Stack Overflow**.
+* In function `backup_object()`, the message will be done after successful saving and not before.
 
 ### Bugs Fixes
-* Function `add_synonym` was not properly working for incomplete entries (missing variables in the replacement values.)
-* Function `load_last` was not properly working for values of `file` without mention of subfolder.
+* Function `add_synonym()` was not properly working for incomplete entries (missing variables in the replacement values.)
+* Function `load_last()` was not properly working for values of `file` without mention of subfolder.
+* Function `accepted_name()` with option `show_traits=TRUE` was not displaying taxa with no entries for taxon traits.
 
 taxlist 0.1.5
 =============
