### 2018-10-27 8cda549
New argument 'accepted_only' in function 'match_names'
