### 2018-10-31 #233 @sckott
@kamapu We've decided to proceed. We do feel it is important though to have functions to convert between the two packages major data structures/classes though, in both packages if possible.  your editor @maelle will follow up with more information.
