### 2018-11-02 #233 @maelle
:wave: @kamapu! I think it'd be great to add the conversion functions before the onboarding process. I'll put this submission on hold, but in the meantime, feel free to ask any questions.
