### 2018-11-02 #233 @maelle
No problem, I might ping you once in a while but that's fine. Please do join the Slack as soon as you want (I sent you an invite).
