+++ DESCRIPTION
@@ -1,10 +1,10 @@
 Package: taxlist
 Version: 0.1.6
 Encoding: UTF-8
-Date: 2018-06-30
 Title: Handling Taxonomic Lists
 Authors@R:
-	person("Miguel", "Alvarez", email="kamapu78@gmail.com", role=c("aut", "cre"))
+	person("Miguel", "Alvarez", email="kamapu78@gmail.com",
+		role=c("aut","cre"), comment=c(ORCID="0000-0003-1500-1834"))
 Depends:
     R(>= 3.0.0),
     ape,
