+++ R/backup_object.R
@@ -34,5 +34,5 @@ backup_object <- function(..., objects=character(), file, stamp=TRUE,
         stamp <- paste(stamp, i, sep="_")
     }
 	save(..., list=objects, file=paste0(file, stamp, ".rda"))
-	message(paste0("Saving backup as '", file, stamp, ".rda'"))
+	message(paste0("Backup saved as '", file, stamp, ".rda'"))
 }
