### 2019-01-10 bd49976
Bug in load_last solved
