### 2019-01-17 71e027d
Function as.phylo moved to own branch
