### 2019-01-20 5da5364
new function 'count_taxa'
