+++ DESCRIPTION
@@ -37,13 +37,12 @@ Collate:
     'NULLing.R''auxiliary_functions.R''deprecated-functions.R'
     'dissect_name.R''clean_strings.R'
     'taxlist-class.R''clean.R''as.list.R''taxon_views.R''add_view.R'
-    'taxon_names.R''taxon_relations.R''taxon_traits.R'
+    'count_taxa.R''taxon_names.R''taxon_relations.R''taxon_traits.R'
     'levels.R''add_concept.R''update_concept.R''add_synonym.R'
     'accepted_name.R''synonyms.R''basionym.R''update_name.R''delete_name.R'
     'replace_view.R''get_children.R'
     'change_concept.R''Extract.R''subset.R'
     'merge_taxa.R''backup_object.R''load_last.R''summary.R'
     'df2taxlist.R''tv2taxlist.R''tnrs.R''tax2traits.R''match_names.R''print_name.R'
-    'as.multiPhylo.R'
     'StartMessage.R'
 VignetteBuilder: knitr
