### 2019-01-21 62e9aae
Minor changes in count_taxa (check for level values)
