### 2019-01-22 kamapu/taxlist#2 @kamapu
Linking 'taxlist' and 'taxa'
Dear @zachary-foster
I'll need your support in this matter. I started with a first attempt to export an object of class `taxlist` to one in `taxa` (I'm not sure now which one, but I'm browsing your vignette from the beginning on).
Here an script creating the example object and handlling it:

```r
library(taxlist)
library(taxa)

# A subset of the example in the package
plants <- subset(Easplist, TaxonName %in% c("Crabbea","Monechma","Pentodon",
				"Commiphora"), keep_parents=TRUE, keep_children=TRUE)

# A table with the accepted names (taxa  does not support synonyms)
plant_table <- accepted_name(plants)
plant_table$Parent <- with(plants@taxonRelations,
		Parent[match(plant_table$TaxonConceptID, TaxonConceptID)])

# Ladies and gentlemen: taxa
plant_taxon <- taxon(
		name=taxon_name(plant_table$TaxonName),
		rank=taxon_rank(paste(plant_table$Level)),
		id=taxon_id(plant_table$TaxonConceptID),
		authority=plant_table$AuthorName)
```
And then, I got a brain jam. My next step would be to use the function `hierarchy()` to establish parent-child relationships but the example in the documentation seems to be much more simple than this case. In `plant_table` I included the information on parent taxa for genus and lower ranks. Can I use it in one command for establishing the hierarchy in object `plant_taxon`? How?
