### 2019-01-24 kamapu/taxlist#2 @kamapu
I continued with the vignette and am, in my opinion, one step forward. Though this may not be the most elegant way to do it, but I can construct parental lines as follows:

```r
plant_hierarchy <- list()
plant_hierarchy[[1]] <- plant_table$TaxonConceptID
i <- 2
repeat{
	plant_hierarchy[[i]] <- plant_table$Parent[match(plant_hierarchy[[i - 1]],
					plant_table$TaxonConceptID)]
	i <- i + 1
	if(!all(is.na(plant_hierarchy[[i - 1]]))) next else break
}
plant_hierarchy <- do.call(cbind, plant_hierarchy)[,-length(plant_hierarchy)]
```

I can even go farther and produce a list with those hierarchies (discarding the top ranks).

```r
plant_hierarchy <- split(plant_hierarchy, 1:nrow(plant_hierarchy))
plant_hierarchy <- lapply(plant_hierarchy, function(x) x[!is.na(x)])
plant_hierarchy <- plant_hierarchy[sapply(plant_hierarchy, function(x) length(x) > 1)]
```

Now is my question, can I use those vectors (which are the taxon identifiers) as indexes in `plant_taxon` to construct the hierarchies? How?

Perhaps I should also include @sckott in this discussion.
