### 2019-01-25 kamapu/taxlist#2 @zachary-foster
Hi @kamapu, I will look at this more soon, but I  think we will be able to do this using `parse_tax_data` from `taxa`
