### 2019-01-25 kamapu/taxlist#2 @zachary-foster
Hi @kamapu, sorry for the delay. I meant `parse_edge_list` in the above comment. I made a rough function to convert from `taxlist` to `taxmap`. It seems to capture most of the information, but might need some work. Hopefully it works on other `taxlist` objects too. It needs the dev version on `taxa` on the `eval` branch:

```r
# devtools::install_github("ropensci/taxa", ref = "eval")

library(taxlist)
library(taxa)

parse_taxlist <- function(intput) {
  # Use the edge list to start making the intput object
  obj = taxa:::parse_edge_list(intput@taxonRelations, taxon_id = "TaxonConceptID", supertaxon_id = "Parent", taxon_name = "TaxonConceptID", taxon_rank = "Level")
  names(obj$data) <- c("relations")

  # Set taxon names and authorities
  concept_data <- intput@taxonNames[! duplicated(intput@taxonNames$TaxonConceptID), ]
  obj$set_taxon_names(concept_data$TaxonName[match(obj$taxon_ids(), concept_data$TaxonConceptID)])
  obj$set_taxon_auths(concept_data$AuthorName[match(obj$taxon_ids(), concept_data$TaxonConceptID)])

  # Add traits table to the intputect
  obj$data$traits <- intput@taxonTraits
  names(obj$data$traits)[1] <- "taxon_id"

  # Add views table to the intputect
  obj$data$views <- intput@taxonViews

  # Add synonyms in a table
  obj$data$synonyms <- intput@taxonNames[duplicated(intput@taxonNames$TaxonConceptID), c("TaxonConceptID", "TaxonName", "AuthorName")]
  names(obj$data$synonyms) <- c("taxon_id", "synonym", "synonym_authority")

  return(obj)
}

obj <- parse_taxlist(Easplist)
obj
```

```
> obj
<Taxmap>
  3887 taxa: 1. Abelmoschus esculentus, 2. Abutilon indicum, 3. Abutilon mauritianum ... 56134. Canellaceae, 56135. Penaeaceae, 56136. Marantaceae
  3887 edges: 54753->1, 54754->2, 54754->3, 54755->4, 54755->5, 54755->6, 54755->7 ... NA->56130, NA->56131, NA->56132, NA->56133, NA->56134, NA->56135, NA->56136
  4 data sets:
    relations:
      # A data.frame: 3887 x 8 (first 3 rows shown)
        taxon_id TaxonConceptID AcceptedName Basionym Parent   Level ViewID  uri
      1        1              1            1       NA  54753 species      1 <NA>
      2        2              2            2       NA  54754 species      1 <NA>
      3        3              3            3       NA  54754 species      1 <NA>
    traits:
      # A data.frame: 311 x 2 (first 3 rows shown)
        taxon_id       lf_behn_2018
      3        7       phanerophyte
      4        9       phanerophyte
      7       18 facultative_annual
    views:
      # A data.frame: 3 x 3
        ViewID                                 secundum view_bibtexkey
      1      1            African Plant Database (2012)  CJBGSANBI2012
      2      2 Taxonomic Name Resolution Service (2018)       TNRS2018
      3      3                    The Plant List (2013)        TPL2013
    synonyms:
      # A data.frame: 1509 x 3 (first 3 rows shown)
         taxon_id               synonym       synonym_authority
      2         1   Hibiscus esculentus                      L.
      5         3        Pavonia patens        (Andrews) Chiov.
      23       20 Spilanthes mauritiana (A. Rich. ex Pers.) DC.
  0 functions:
```
