### 2019-01-25 kamapu/taxlist#2 @kamapu
@zachary-foster
This will get funny (it is a quite different logic than mine but I will trust rather your way). Just few comments:
1) Could you commit the function in a fork, so we can also document it evolution.
2) I think the assignation of accepted names and synonyms through function `duplicated()` can be risky: The accepted name of a taxon may also appear after its synonyms in the slot `taxonNames`.
3) I was wrong, you seem also to support synonyms in `taxa`.
4) The final function may also be set as S4 method (I have the impresion, `vegtable` objects may also be converted into `taxmap` ones).
I could care on 2) and 4), once the function is in GitHub.
