### 2019-01-26 fb00a17
function 'match_names()' improved
