+++ man/match_names.Rd
@@ -33,6 +33,13 @@ using the function \code{\link{stringsim}}.
 For \code{output="list"} a list with the best matches (taxon usage name ID and
 similarity) for each queried name will be retrieved, where the number is set by
 argument \code{'best'}.
+
+Option \code{'accepted_only=TRUE'} will only work with
+\code{'output="data.frame"'}.
+This will be applied especially in those cases were the requested names have
+more than one match in the reference \code{\linkS4class{taxlist}} object
+(matching homonyms) and will retrieve the one name, that has the status of
+accepted name, otherwise no matchings will be retrieved.
 }
 \author{
 Miguel Alvarez, \email{kamapu78@gmail.com}.
