+++ NEWS.md
@@ -1,3 +1,11 @@
+taxlist 0.1.7
+=============
+
+### Bug Fixes
+
+* Function `match_names()` was not properly working for the option `accepted_only=TRUE`.
+
+
 taxlist 0.1.6
 =============
 
