+++ DESCRIPTION
@@ -1,5 +1,5 @@
 Package: taxlist
-Version: 0.1.6
+Version: 0.1.7
 Encoding: UTF-8
 Title: Handling Taxonomic Lists
 Authors@R:
