### 2019-01-29 kamapu/taxlist#2 @zachary-foster
>Could you commit the function in a fork, so we can also document it evolution?

Its on the "eval" branch of taxa. Do you mean on a fork of `taxlist`? I am fine with it being in either package. `taxa` does not have parsers for specific formats (since its a general-use foundational package), so it might make more sense to have it in `taxlist`.

> I was wrong, you seem also to support synonyms in taxa.

In a way, but not explicitly. `taxa` supports arbitrary data assigned to taxa. The synonym table is just an arbitrary table with rows assigned to taxa (all the other tables are too). So, the `taxamap` object does not "know" it contains synonyms, if that makes sense.

> I could care on 2) and 4), once the function is in GitHub.

Sounds good. Let me know if you need help with manipulating `taxmap` objects and such. You can name the tables and columns whatever you want, but each table needs a "taxon_id" column if it contains per-taxon data, otherwise the filtering functions will not work with that table.
