### 2019-01-29 kamapu/taxlist#2 @kamapu
OK. If you don't mind @zachary-foster I will just copy the code and adapt it to a function in `taxlist` and will try to do the *counter-function*, though I guess, it won't be possible to do transformations from `taxlist` to `taxamap` and back without loosing information.
