### 2019-01-29 kamapu/taxlist#2 @zachary-foster
> OK. If you don't mind @zachary-foster I will just copy the code and adapt it to a function in taxlist

That is fine with me

>  it won't be possible to do transformations from taxlist to taxamap and back without loosing information.

Perhaps not with the function I wrote, but it should be possible.
