+++ TODO.md
@@ -1,5 +1,11 @@
-* Improve documentation in `print_name`
-* Solve problems with `as.phylo` (communication with developer of package `ape`)
-* Check for function `tnrs`
-* Make function `check_names` more efficient for big species lists
-* In `match_names` an argument to compare only among accepted names
+### TODO
+
+[] Change the default name (**TaxonConceptID**) in output of `count_taxa()`.
+[] Improve documentation in `print_name()` and test different options.
+[] Solve problems with `as.phylo()` (communication with developer of package `ape`).
+[] Check for function `tnrs()` and improve it.
+[] Make function `check_names()` more efficient for big species lists.
+
+### Achieved
+
+[x] An argument in `match_names()` to compare only among accepted names.
