+++ TODO.md
@@ -1,9 +1,13 @@
 ### TODO
 
 [] Change the default name (**TaxonConceptID**) in output of `count_taxa()`.
+
 [] Improve documentation in `print_name()` and test different options.
+
 [] Solve problems with `as.phylo()` (communication with developer of package `ape`).
+
 [] Check for function `tnrs()` and improve it.
+
 [] Make function `check_names()` more efficient for big species lists.
 
 ### Achieved
