### 2019-02-23 d34ebef
New function 'replace_x' and allies
