### 2019-02-25 a66fed3
Small changes in summary
