### 2019-05-08 bf1deb7
New functions 'taxlist2taxmap' and 'taxmap2taxlist'
