+++ R/taxmap2taxlist.R
@@ -1,0 +1,102 @@
+# TODO:   Parse taxmap objects to taxlist
+# 
+# Author: Miguel Alvarez
+################################################################################
+
+taxmap2taxlist <- function(taxmap, relations, traits, synonyms, views,
+		reindex=FALSE) {
+	# Extract concept IDs
+	old_ids <- taxon_ids(taxmap)
+	if(reindex)
+		TaxonConceptID <- as.integer(1:length(old_ids)) else
+		TaxonConceptID <- as.integer(old_ids)
+	if(any(is.na(TaxonConceptID)))
+		stop("Some values of 'taxon_id' cannot be coerced to integers. Try 'reindex=TRUE'")
+	
+	# Extract taxon relations
+	if(missing(relations)) {
+		taxonRelations <- data.frame(
+				TaxonConceptID=TaxonConceptID,
+				TaxonName=taxon_names(taxmap),
+				stringsAsFactors=FALSE)
+	} else {
+		taxonRelations <- taxmap$data[[relations]]
+		colnames(taxonRelations)[colnames(taxonRelations) ==
+						"taxon_id"] <- "TaxonConceptID"
+		taxonRelations$TaxonConceptID <- TaxonConceptID[
+				match(taxonRelations$TaxonConceptID, old_ids)]
+	}
+	
+	# Extract names IDs and respective information
+	if(!"TaxonUsageID" %in% colnames(taxonRelations))
+		taxonRelations$TaxonUsageID <- as.integer(1:nrow(taxonRelations))
+	if(!"TaxonName" %in% colnames(taxonRelations))
+		taxonRelations$TaxonName <- taxon_names(taxmap)
+	if(!"AuthorName" %in% colnames(taxonRelations) &
+			!is.null(taxmap$authorities))
+		taxonRelations$AuthorName <- taxmap$authorities
+	if(!"AuthorName" %in% colnames(taxonRelations) &
+			is.null(taxmap$authorities))
+		taxonRelations$AuthorName <- as.character(NA)
+	taxonNames <- taxonRelations[,c("TaxonUsageID", "TaxonConceptID",
+					"TaxonName", "AuthorName")]
+	taxonNames$AuthorName[taxonNames$AuthorName == "NA"] <- NA
+	
+	# Completing taxonRelations
+	taxonRelations$AcceptedName <- taxonRelations$TaxonUsageID
+	taxonRelations <- taxonRelations[,!colnames(taxonRelations) %in%
+					c("TaxonUsageID", "TaxonName", "AuthorName")]
+	taxonRelations$Level <- taxon_ranks(taxmap)
+	taxonRelations$Parent <- taxmap$edge_list$from
+	taxonRelations$Parent <- TaxonConceptID[match(taxonRelations$Parent,
+					old_ids)]
+	if(!"Basionym" %in% colnames(taxonRelations))
+		taxonRelations$Basionym <- as.integer(NA)
+	if(!"ViewID" %in% colnames(taxonRelations))
+		taxonRelations$ViewID <- as.integer(NA)
+	
+	# Adding synonyms to taxonNames
+	if(!missing(synonyms)) {
+		Synonyms <- taxmap$data[[synonyms]]
+		colnames(Synonyms) <- taxlist::replace_x(colnames(Synonyms),
+				c("taxon_id", "synonym", "synonym_authority"),
+				c("TaxonConceptID", "TaxonName", "AuthorName"))
+		if(!"TaxonUsageID" %in% colnames(Synonyms))
+			Synonyms$TaxonUsageID <- as.integer(max(taxonNames$TaxonUsageID) +
[…]
