+++ NAMESPACE
@@ -38,6 +38,7 @@ exportMethods(
 	summary,
 	synonyms,
 	tax2traits,
+	taxlist2taxmap,
 	taxon_names,"taxon_names<-",
 	taxon_relations,"taxon_relations<-",
 	taxon_traits,"taxon_traits<-",
@@ -61,5 +62,6 @@ export(
 	replace_na,
 	replace_x,
 	S4_to_list,
+	taxmap2taxlist,
 	tv2taxlist
 )

[…]
