### 2019-05-08 d8823a0
Correction in .Rbuildignore
