### 2019-05-09 kamapu/taxlist#2 @kamapu
Hello @zachary-foster
I just came back from another trip and started working on the functions and now I have one. I adapted your code and wrote one function to convert objects from `taxlist` to `Taxmap`, and wrote a second function for the conversion back. You will need then to install `taxlist` from my branc:
```r
library(devtools)
install_github("kamapu/taxlist", ref="miguel")
install_github("ropensci/taxa", ref="eval")
```
I included an R-image with three examples:
* `data1` a subset of `Easplist`
* `data2` the same subset but only species and without taxonomy, only synonyms
* `data3` the previous example without synonyms
```r
library(taxlist)
library(taxa)
load(file.path(path.package("taxlist"), "taxlist_examples/examples.Rda"))
```
Export to `taxa` is done by `taxlist2taxmap()`:
```r
Test1 <- taxlist2taxmap(data1)
Test2 <- taxlist2taxmap(data2)
Test3 <- taxlist2taxmap(data3)
```
For the export back, there is the function `taxmap2taxlist()`:
```
Test4 <- taxmap2taxlist(Test1, relations="relations", traits="traits", synonyms="synonyms",
    views="views")
summary(Test4)

Test5 <- taxmap2taxlist(ex_taxmap, traits="info", reindex=TRUE)
summary(Test5)
```
Note that the later is working with the example from package `taxa`.
