### 2019-05-09 kamapu/taxlist#2 @kamapu
Some comments to my later message:

1. The example will only work with the branch **eval** of `taxa`. Are you planing to merge it with **master**?
2. If I apply some functions in these `Taxmap` objects (e.g. `print_tree()`) I will get a warning regarding the table **views** which is missing the column **taxon_id**.
3. I had the impression, the function `print_tree()` is not working properly at the moment. I only get names of families and NAs in the display, where I was expecting a tree.
