### 2019-05-09 #233 @kamapu
Hello @maelle and @noamross
There is one recent step [here](https://github.com/kamapu/taxlist/issues/2#issuecomment-490748989)
We are working in a function for converting `taxlist` objects into `Taxmap`.
