### 2019-05-10 kamapu/taxlist#2 @zachary-foster
Hi @kamapu

Cool, I hope you had a nice trip!

> I included an R-image with three examples

Is the R image in the package or was it supposed to be attached to this issue? I don't see it.

> The example will only work with the branch eval of taxa. Are you planing to merge it with master?

Yes, it will be merged eventually. I am hoping to get a change to work on it soon.

>If I apply some functions in these Taxmap objects (e.g. print_tree()) I will get a warning regarding the table views which is missing the column taxon_id.

Thats ok, I think I will remove that warning eventually. It happens when you filter taxa but some of the data is not classified by taxa so it cant be filtered.

> I had the impression, the function print_tree() is not working properly at the moment. I only get names of families and NAs in the display, where I was expecting a tree.

It should be working, but I have seen that error once before, but have never been able to reproduce it. If I can reproduce it with the code you supplied, perhaps I can finally fix it!

> I realized that the size of Taxmap objects is much smaller than the taxlist ones, thought the content is almost the same. Is this inherent to R6 or I'm doing something wrong?

You mean how much RAM it takes? I am not sure. Its probably a difference between how S4 and R6 stores data. R6 is pretty lightweight as I understand it. I don't have much experience with S4. I will look into it when I test out the code.
