### 2019-05-13 kamapu/taxlist#2 @kamapu
Hi @zachary-foster : Here a quick reply.

> Is the R image in the package or was it supposed to be attached to this issue? I don't see it.

The image will be installed with `taxlist` from the **miguel** branch.

```r
library(devtools)
install_github("kamapu/taxlist", ref="miguel")
```

Then you just need following command to get the objects in your session:

```r
library(taxlist)
load(file.path(path.package("taxlist"), "taxlist_examples/examples.Rda"))
```

Note, this image is not in **master**.

> You mean how much RAM it takes? [...] R6 is pretty lightweight as I understand it [...]

Yes, I meant the allocated RAM. If you compare outputs of `object.size()` once the objects are converted, you will wander, how big the differences are... unfear!

Regarding the warning message, I assume, it is OK, eventhough some users get scared from warnings (they confuse them with error messages) and it can cause some troubles when checking examples.

Just to remind you, the export/import functions were requested by the editor to proceed with the submission of `taxlist` to **ROpenSci**. I assume, we may have to get it working on the respective **master** branches, before I request to continue with the process.
