### 2019-05-13 kamapu/taxlist#2 @zachary-foster
Thanks for the info on finding the test data! I did not see it in your first message until now.

These function seem to work fine for me, at least with the test data.

I cannot replicate the `print_tree` NA bug. They all look normal for me.

In regards to memory use, I recommend using `pryr::object_size`, since R6 objects are environments and `object.size` does not count environments.  `taxlist` objects are actually much smaller than `taxmap` objects, since R6 objects store all of the functions that operate on them with the object and S4 don't. I would guess that both will have about the same RAM requirements for very large objects.

```
> object_size(Test1$data)
19.5 kB
> object_size(data1)
25.2 kB
> object_size(Test1)
1.62 MB
```

>  I assume, we may have to get it working on the respective master branches, before I request to continue with the process.

Ok, i will try to get eval pushed to master soon.
