### 2019-06-06 bae6049
Fixed bug in merge_taxa()
