+++ R/merge_taxa.R
@@ -21,7 +21,10 @@ setMethod("merge_taxa", signature(object="taxlist", concepts="numeric",
 			if(any(!concepts %in% object@taxonRelations$TaxonConceptID))
 				stop("All values in 'concepts' should be included as concepts in 'object'")
 			# Merging concepts
-			object@taxonNames[object@taxonNames$TaxonConceptID %in% concepts, "TaxonConceptID"] <- concepts[1]
+			object@taxonNames[object@taxonNames$TaxonConceptID %in% concepts,
+					"TaxonConceptID"] <- concepts[1]
+			object@taxonRelations[object@taxonRelations$Parent %in% concepts,
+					"Parent"] <- concepts[1]
 			object@taxonRelations <- object@taxonRelations[
 					!object@taxonRelations$TaxonConceptID %in% concepts[-1],]
 			object <- clean(object)
