### 2019-07-02 kamapu/taxlist#2 @zachary-foster
I am currently rewriting the fundamental classes of `taxa` as S3 vectors using the `vctrs` package so they act more like base R, so much of the changes in `eval` will probably not be used unfortunately.  I will try to figure out what part of `eval` you need to get the conversion functions to work and merge that part.
