### 2019-07-04 kamapu/taxlist#2 @kamapu
Great! I tested it and got no errors!
Then, I will proceed to ROpenSci...
