### 2019-07-07 a0dd760
Warning message in tax2traits for objects without taxonomic rank
