+++ NEWS.md
@@ -6,6 +6,10 @@ taxlist 0.1.7
 * Set of functions for data manipulation, namely `replace_x()`, `replace_idx()`, `replace_na()`, and `insert_rows()`.
 * Exchange of data with the package **taxa** through the functions `taxlist2taxmap()` and `taxmap2taxlist()`.
 
+### Improvements
+
+* Warning in function `tax2traits()` for objects without taxonomic ranks.
+
 ### Bug Fixes
 
 * Function `match_names()` was not properly working for the option `accepted_only=TRUE`.
