### 2019-07-15 #233 @kamapu
Dear @noamross Do you mean, I should follow [this instructions?](http://r-pkgs.had.co.nz/tests.html)
I'm a self-made programmer and now I'm feeling like someone, who is driving without driving license
