### 2019-07-15 #233 @noamross
> I'm a self-made programmer and now I'm feeling like someone, who is driving without driving license
As are most of us!  Yes, that's where to start for writing tests.
