### 2019-10-06 79df543
First attempt to implement code coverage
