### 2019-10-08 7f01e09
Bug in clean solved
