+++ codecov.yml
@@ -1,0 +1,19 @@
+comment: false
+language: R
+sudo: false
+cache: packages
+after_success:
+  - Rscript -e 'covr::codecov()'
+
+comment: false
+
+coverage:
+  status:
+    project:
+      default:
+        target: auto
+        threshold: 1%
+    patch:
+      default:
+        target: auto
+        threshold: 1%
