+++ tests/testthat/test-levels.R
@@ -1,0 +1,12 @@
+# TODO:   Add comment
+# 
+# Author: Miguel Alvarez
+################################################################################
+
+context("retrieving levels")
+
+test_that("levels in taxlist object", {
+			expect_equal(levels(Easplist),
+					base::levels(Easplist@taxonRelations$Level))
+		}
+)
