+++ tests/testthat/test-as_list.R
@@ -1,0 +1,11 @@
+# TODO:   Add comment
+# 
+# Author: Miguel Alvarez
+################################################################################
+
+context("coercion to list")
+
+test_that("coercion to list works", {
+			expect_is(as.list(Easplist), "list")
+		}
+)
