+++ .travis.yml
@@ -12,4 +12,4 @@ r:
 r_check_args: "--as-cran"
 
 after_success:
-  - Rscript -e 'covr::codecov()'
+  - Rscript -e 'covr::codecov(token="61b4bc87-9f95-4b70-b0ad-0bf8bce07df1")'
