+++ tests/testthat/test-tax2traits.R
@@ -1,0 +1,12 @@
+# TODO:   Add comment
+# 
+# Author: Miguel Alvarez
+################################################################################
+
+context("inserting taxonomic information in slot taxonTraits")
+
+test_that("comparing columns in taxon traits", {
+			expect_equal(rncol(Easplist@taxonTraits) <
+							ncol(tax2traits(Easplist)@taxonTraits), TRUE)
+		}
+)
