+++ tests/testthat/test-tax2traits.R
@@ -6,7 +6,7 @@
 context("inserting taxonomic information in slot taxonTraits")
 
 test_that("comparing columns in taxon traits", {
-			expect_equal(rncol(Easplist@taxonTraits) <
+			expect_equal(ncol(Easplist@taxonTraits) <
 							ncol(tax2traits(Easplist)@taxonTraits), TRUE)
 		}
 )
