+++ tests/testthat/test-taxon_traits.R
@@ -1,0 +1,11 @@
+# TODO:   Add comment
+# 
+# Author: Miguel Alvarez
+################################################################################
+
+context("manipulating taxon usage names")
+
+test_that("function taxon_names is working", {
+			expect_is(taxon_traits(Easplist), "data.frame")
+		}
+)
