+++ tests/testthat/test-taxon_traits.R
@@ -3,7 +3,7 @@
 # Author: Miguel Alvarez
 ################################################################################
 
-context("manipulating taxon usage names")
+context("manipulating taxon traits")
 
 test_that("function taxon_names is working", {
 			expect_is(taxon_traits(Easplist), "data.frame")
