### 2019-10-13 2e2cb16
Completing test for Extract.R
