### 2019-10-13 6e0dea4
Completing code coverage until 'add_view'
