+++ tests/testthat/test-count_taxa.R
@@ -5,8 +5,12 @@
 
 context("counting taxa")
 
-test_that("count taxa is working", {
+test_that("count_taxa is working", {
 			expect_equal(count_taxa(iris$Species), length(unique(iris$Species)))
 			expect_equal(count_taxa(Easplist), nrow(Easplist@taxonRelations))
+			expect_equal(count_taxa(Easplist, "species"),
+					with(Easplist@taxonRelations,
+							length(Level[paste(Level) == "species"])))
+			expect_error(count_taxa(Easplist, "superspecies"))
 		}
 )
