+++ tests/testthat/test-print_name.R
@@ -10,6 +10,13 @@ test_that("printed names are working", {
 					"expression")
 			expect_is(print_name(Easplist, 363, style="expression",
 							secundum="secundum"), "expression")
+			## expect_is(print_name(Easplist, 363), "AsIs")
+			expect_equal(paste(print_name(Easplist, 363)),
+					"*Ludwigia adscendens* ssp. *diffusa* (Forssk.) P.H. Raven")
+			expect_equal(paste(print_name(Easplist, 363, style="html")),
+					"<i>Ludwigia adscendens</i> ssp. <i>diffusa</i> (Forssk.) P.H. Raven")
+			expect_equal(paste(print_name(Easplist, 363, style="knitr")),
+					"\\textit{Ludwigia adscendens} ssp. \\textit{diffusa} (Forssk.) P.H. Raven")
 			## expect_equal(print_name(Easplist, 363),
 			##         "*Ludwigia adscendens* ssp. *diffusa* (Forssk.) P.H. Raven")
 			## expect_equal(print_name(Easplist, 363, style="html"),
