+++ tests/testthat/test-tv2taxlist.R
@@ -1,0 +1,12 @@
+# TODO:   Add comment
+# 
+# Author: Miguel Alvarez
+################################################################################
+
+context("importing Turboveg data sets")
+
+test_that("import from Turboveg is working", {
+			expect_is(tv2taxlist("cyperus", file.path(path.package("taxlist"),
+									"tv_data")), "taxlist")
+		}
+)
