### 2019-10-18 b686c28
solving bug in testthat
