+++ tests/testthat/test-taxlist2taxmap.R
@@ -11,8 +11,8 @@ data1 <- taxlist2taxmap(data1)
 test_that("conversion to taxa is working", {
 			expect_equal(class(taxlist2taxmap(Easplist)),
 					c("Taxmap", "Taxonomy", "R6"))
-			expect_equal(class(taxmap2taxlist(data1, relations="relations",
-									traits="traits", synonyms="synonyms", 
-									views="views")), "taxlist")
+			expect_is(taxmap2taxlist(data1, relations="relations",
+							traits="traits", synonyms="synonyms",
+							views="views"), "taxlist")
 		}
 )
