### 2019-10-18 #233 @kamapu
@maelle I managed to implement the test suite. At the moment `taxlist` has the requested 75% of code coverage. I'll try to improve it but I though, we can proceed with the submission process. I hope, this was the hardest part (everything was new for me).
