### 2019-10-18 kamapu/taxlist#2 @kamapu
@zachary-foster Is there a release of the current version of `taxa` planned for soon? I ask it because the conversions from `taxlist` objects and back is only working with the last GitHub version but not with the CRAN one, thus a submission to CRAN won't work at this stage.
