### 2019-10-18 kamapu/taxlist#2 @zachary-foster
I had not planned one, but I probably could release an updated version to CRAN if you need. I am about 70% done rewriting `taxa` to make the more basic classes like base R vectors (can use as columns in tables and such), so I have not worked on the `master` branch for a while.
