### 2019-10-18 #233 @maelle
:wave: @kamapu! Great news! Congrats on adding the tests, I imagine it wasn't easy but you've learnt something very useful. :smile_cat:

### Editor checks:

---

#### Editor comments

* A big ask I have before looking for reviewers would be a summary of https://github.com/ropensci/taxa/issues/130 in the form of a README section aimed at helping potential users choose between taxlist and taxa depending on their use case/coding style preferences. :wink: It'd be ideal if taxa then featured the same information cc @zachary-foster @sckott :-)

* Is taxlist aimed at being mainly developer facing i.e. used as a dependency of other packages?

* I'm getting an error when running the tests.

```r
── 1. Error: (unknown) (@test-taxlist2taxmap.R
attempt to apply non-function
1: taxlist2taxmap(data1) at testthat/test-taxlist2taxmap.R:9
2: taxlist2taxmap(data1)

══ testthat results  ═════════════════════════
[ OK: 91 | SKIPPED: 0 | WARNINGS: 0 | FAILED: 1 ]
1. Error: (unknown) (@test-taxlist2taxmap.R#9)
```

* `goodpractice` output

```r
 ── GP taxlist ─────────────────────────────────────────────────────

 It is good practice to

 ✖ not use "Depends" in DESCRIPTION, as it can
 cause name clashes, and poor interaction with other
 packages. Use "Imports" instead.
```
```r

 ✖ avoid long code lines, it is bad for
 readability. Also, many people prefer editor windows
 that are about 80 characters wide. Try make your lines
 shorter than 80 characters

 R/accepted_name.R:54:1
 R/add_synonym.R:16:1
 R/backup_object.R:32:1
 R/basionym.R:48:1
 R/clean.R:43:1
 ... and 36 more lines

 ✖ avoid sapply(), it is not type safe. It might
 return a vector, or a list, depending on the input
 data. Consider using vapply() instead.

 R/dissect_name.R:8:9
 R/dissect_name.R:10:9

 ✖ avoid 1:length(...), 1:nrow(...),
 1:ncol(...), 1:NROW(...) and 1:NCOL(...) expressions.
 They are error prone and result 1:0 if the expression
 on the right hand side is zero. Use seq_len() or
 seq_along() instead.

 R/add_concept.R:73:42
 R/df2taxlist.R:76:36
 R/df2taxlist.R:76:64
 R/load_last.R:35:17
 R/match_names.R:20:27
 ... and 7 more lines

 ✖ not import packages as a whole, as this can
 cause name clashes between the imported packages.
 Instead, import only the specific functions you need.

```

A few things to fix, do not hesitate to ask for help via this thread/discuss.ropensci.org!

* Once the previous items are fixed and once reviewers are found the master branch of the repo should remain stable (well apart from bug fixes of course) so that the reviewers might be looking at something stable rather than a moving target (writing this because of your remark regarding improving the tests).
---

Reviewers:
Due date:
