### 2019-11-10 1e4d579
Fulfilling good practices
