+++ R/dissect_name.R
@@ -5,8 +5,8 @@
 
 dissect_name <- function(x, split=" ", fixed=TRUE, ...) {
 	x <- strsplit(x, split=split, fixed=fixed, ...)
-	LEN <- sapply(x, length)
+	LEN <- unlist(lapply(x, length))
 	Expand <- function(y, z) c(y, rep(NA, z - length(y)))
-	x <- t(sapply(x, Expand, z=max(LEN)))
+	x <- do.call(rbind, lapply(x, Expand, z=max(LEN)))
 	return(x)
 }
