+++ NAMESPACE
@@ -1,14 +1,12 @@
 ## Import functions
-importFrom(stats, aggregate, na.omit, sd)
+importFrom(methods, new, slot, slotNames, validObject)
 importFrom(utils, object.size)
+importFrom(stats, aggregate, na.omit, sd)
 importFrom(vegdata, tv.home)
 importFrom(taxize, tnrs)
 importFrom(foreign, read.dbf)
 importFrom(stringdist, stringsim)
 importFrom(taxa, taxon_ids, taxon_ranks)
-
-## Import of packages
-import(methods)
 
 ## Export of classes
 exportClasses(taxlist)
