### 2019-11-10 #233 @kamapu
Dear @maelle you are right, I have learnt a lot and got some new skills for programming packages.

Although I still have to do the comparison with `taxa` in README, I respond herewith to the rest of your comments in advance.

> - Is taxlist aimed at being mainly developer facing i.e. used as a dependency of other packages?

Not and yes. The package `taxlist` is handling taxonomic information within objects defined by the package `vegtable`, the later handling biodiversity records. Nevertheless the users need to know objects and functions defined in `taxlist` to exploit the capabilities of both packages.

> - I'm getting an error when running the tests.

This is due to the function `taxlist2taxmap()`, which is depending on the current version of `taxa` at **GitHub** but is not working with the last version at **CRAN**. I requested @zachary-foster to update the CRAN version (see [here](https://github.com/kamapu/taxlist/issues/2#issuecomment-543629925)). As soon as `taxa` gets updated, this error should disappear.

> - `goodpractice` output

All issues have been solved with exception of the second one ("avoid long code lines, it is bad for
 readability"). I am also working with an editor of 80 characters wide and use to break code for readability. The problem is caused by strings included in `stop()` and `warning()` statements, which I prefer not to break. Since those lines use to be within functions and therefore indented, the width limit is exceeded. I hope this issue can be tolerated, otherwise I'll need to suppress the indentations in those specific lines.

Required changes have been committed to the master branch.

Pending tasks are therefore:
- Insert comparison with `taxa` in README
- Increase coverage?
