### 2019-11-11 ad3bfcc
adding backup_object to test suite
