+++ tests/testthat/test-taxon_relations.R
@@ -5,6 +5,9 @@
 
 context("manipulating taxon concepts")
 
+# Loading installed examples
+load(file.path(path.package("taxlist"), "taxlist_examples/examples.Rda"))
+
 test_that("function taxon_relations is working", {
 			expect_is(taxon_relations(Easplist), "data.frame")
 			tmp <- Easplist
@@ -33,6 +36,17 @@ test_that("function add_concept is working", {
 			Diospyros <- subset(Easplist, grepl("Diospyros", TaxonName),
 					slot="names", keep_parents=TRUE)
 			expect_error(add_concept(Diospyros, Euclea))
+			# Working with lists without information on taxonomic levels
+			expect_equal(nrow(add_concept(data2,
+									TaxonName="Planta incognita")@taxonRelations) >
+					nrow(data2@taxonRelations), TRUE)
+			expect_warning(add_concept(data2, TaxonName="Planta incognita",
+							Level="species"))
+			# Adding on a new object
+			expect_equal(nrow(add_concept(new("taxlist"),
+									TaxonName="Planta incognita")@taxonRelations),
+					1)
+			
 		}
 )
 
