### 2019-11-11 2217166
Solve error in backup_object
