+++ tests/testthat/test-backup_object.R
@@ -6,7 +6,7 @@
 context("Creating backups")
 
 # Loading installed examples
-load(file.path(path.package("taxlist"), "inst/taxlist_examples/examples.Rda"))
+load("examples.rda")
 
 test_that("Backups of objects are written", {
 			expect_output_file(backup_object(data1, file="backup", stamp=FALSE,
