+++ tests/testthat/test-taxon_relations.R
@@ -6,7 +6,8 @@
 context("manipulating taxon concepts")
 
 # Loading installed examples
-load(file.path(path.package("taxlist"), "taxlist_examples/examples.Rda"))
+## load(file.path(path.package("taxlist"), "taxlist_examples/examples.Rda"))
+load("examples.rda")
 
 test_that("function taxon_relations is working", {
 			expect_is(taxon_relations(Easplist), "data.frame")
