### 2019-11-11 7e314b9
discarded test for backup_object
