### 2019-11-13 #233 @maelle
Thank you for your work and answer!

*
> This is due to the function taxlist2taxmap(), which is depending on the current version of taxa at GitHub but is not working with the last version at CRAN.

In the meantime please use a Remotes field in DESCRIPTION cf https://remotes.r-lib.org/articles/dependencies.html#github

* Regarding the long lines for warnings, add ""# nolint" at the end cf https://github.com/jimhester/lintr#project-configuration

* taxa is mostly aimed at developers, maybe a point for the comparison. cf https://github.com/ropensci/taxa/issues/201#issuecomment-543868257

* Reg test coverage yes it is just at the limit at the moment. :wink:
