### 2019-12-13 302aad5
Possible problems with existing objects
