+++ R/clean.R
@@ -40,17 +40,28 @@ setMethod("clean", signature(object="taxlist"),
 				if(count == times) break
 			}
 			# clean classes of key fields
-			object@taxonNames$TaxonConceptID <- as.integer(object@taxonNames$TaxonConceptID)
-			object@taxonNames$TaxonUsageID <- as.integer(object@taxonNames$TaxonUsageID)
-			object@taxonNames$TaxonName <- paste(object@taxonNames$TaxonName)
-			object@taxonNames$AuthorName <- paste(object@taxonNames$AuthorName)
-			object@taxonRelations$TaxonConceptID <- as.integer(object@taxonRelations$TaxonConceptID)
-			object@taxonRelations$AcceptedName <- as.integer(object@taxonRelations$AcceptedName)
-			object@taxonRelations$Basionym <- as.integer(object@taxonRelations$Basionym)
-			object@taxonRelations$Parent <- as.integer(object@taxonRelations$Parent)
-			object@taxonRelations$ViewID <- as.integer(object@taxonRelations$ViewID)
-			object@taxonTraits$TaxonConceptID <- as.integer(object@taxonTraits$TaxonConceptID)
-			object@taxonViews$ViewID <- as.integer(object@taxonViews$ViewID)
+			object@taxonNames$TaxonConceptID <-
+					as.integer(object@taxonNames$TaxonConceptID)
+			object@taxonNames$TaxonUsageID <-
+					as.integer(object@taxonNames$TaxonUsageID)
+			object@taxonNames$TaxonName <-
+					paste(object@taxonNames$TaxonName)
+			object@taxonNames$AuthorName <-
+					paste(object@taxonNames$AuthorName)
+			object@taxonRelations$TaxonConceptID <-
+					as.integer(object@taxonRelations$TaxonConceptID)
+			object@taxonRelations$AcceptedName <-
+					as.integer(object@taxonRelations$AcceptedName)
+			object@taxonRelations$Basionym <-
+					as.integer(object@taxonRelations$Basionym)
+			object@taxonRelations$Parent <-
+					as.integer(object@taxonRelations$Parent)
+			object@taxonRelations$ViewID <-
+					as.integer(object@taxonRelations$ViewID)
+			object@taxonTraits$TaxonConceptID <-
+					as.integer(object@taxonTraits$TaxonConceptID)
+			object@taxonViews$ViewID <-
+					as.integer(object@taxonViews$ViewID)
 			return(object)
         }
 )
