+++ .travis.yml
@@ -1,9 +1,9 @@
 language: r
 cache: packages
-os:
-  - windows
-  - linux
-  - osx
+#os:
+#  - windows
+#  - linux
+#  - osx
 
 r_github_packages:
   - jimhester/covr
