### 2019-12-18 48a985a
New method for function 'match_names'
