+++ DESCRIPTION
@@ -9,7 +9,7 @@ Authors@R: c(
         role="ctb", comment=c(ORCID = "0000-0002-5075-0948"))
     )
 Depends:
-    R(>= 3.0.0)
+    R(>= 3.5.0)
 Imports:
     foreign,
     methods,
