### 2019-12-26 847952b
New CRAN version
