### 2020-01-10 a7b9f79
Improved coverage and pre-release version
