+++ tests/testthat/test-match_names.R
@@ -7,7 +7,8 @@ context("matching names")
 
 test_that("match_names is working", {
 			expect_is(match_names(c("Prosopis juliflora","Opuntia vulgaris"),
-							Easplist), "data.frame")
+							Easplist, accepted_only=TRUE,
+							show_concepts=TRUE), "data.frame")
 			expect_equal(nrow(match_names(c("Prosopis juliflora",
 											"Opuntia vulgaris"), Easplist)), 2)
 			expect_is(match_names(c("Prosopis juliflora","Opuntia vulgaris"),
@@ -15,5 +16,10 @@ test_that("match_names is working", {
 			expect_equal(length(match_names(c("Prosopis juliflora",
 											"Opuntia vulgaris"), Easplist,
 									output="list")), 2)
+			expect_equal(length(match_names("Poa pratensis",
+									c("Poa annua", "Geranium pratensis"))), 5)
+			expect_warning(match_names(
+							c("Geranium robertianum", "Poa pratensis"),
+							c("Poa annua", "Geranium pratensis")))
 		}
 )
