### 2020-01-10 #233 @kamapu
Happy new year @maelle
I answer late, because I'm still working on it and was waiting to have good news for you:

1) Long code was solved by skipping some space from indentation and using `paste()` for too long messages.
2) Coverage have been also improved a bit, but I am still struggling with some issues on this regard (see questions at the end).
3) Since I have to submit a new version of `taxlist` to CRAN, I had to remove the functions `taxlist2taxmap()` and `taxmap2taxlist()` from master and will reinsert them, once I get news about a new CRAN version of `taxa`, which is compatible with those functions.

I try to finish with the file **README.md** and come back to you. In the meantime some questions:

1) I don't have any clue about how to test writing functions, specifically in the case of `backup_object()`, which is writing a **.rda** file. I was not able to find help on Internet (some advices were not comprehensive for my background).
2) The same is valid for functions printing in the console, for instance `taxlist::summary()`.

Some hints for me?
