+++ README.md
@@ -1,79 +1,43 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
-
 <!-- Use snippet 'render_markdown' for it -->
-
-# taxlist
+taxlist
+=======
 
 <!-- Budges -->
+[![CRAN\_Status\_Badge](http://www.r-pkg.org/badges/version/taxlist)](https://cran.r-project.org/package=taxlist) [![Rdoc](http://www.rdocumentation.org/badges/version/taxlist)](http://www.rdocumentation.org/packages/taxlist) [![DOI](https://zenodo.org/badge/54913161.svg)](https://zenodo.org/badge/latestdoi/54913161) <br> [![Travis Build Status](https://travis-ci.org/kamapu/taxlist.svg?branch=master)](https://travis-ci.org/kamapu/taxlist) [![Codecov test coverage](https://codecov.io/gh/kamapu/taxlist/branch/master/graph/badge.svg)](https://codecov.io/gh/kamapu/taxlist?branch=master) <br> [![CRAN\_downloads](http://cranlogs.r-pkg.org/badges/taxlist)](https://cran.r-project.org/package=taxlist) [![total downloads](http://cranlogs.r-pkg.org/badges/grand-total/taxlist)](https://cran.r-project.org/package=taxlist)
 
-[![CRAN\_Status\_Badge](http://www.r-pkg.org/badges/version/taxlist)](https://cran.r-project.org/package=taxlist)
-[![Rdoc](http://www.rdocumentation.org/badges/version/taxlist)](http://www.rdocumentation.org/packages/taxlist)
-[![DOI](https://zenodo.org/badge/54913161.svg)](https://zenodo.org/badge/latestdoi/54913161)
-<br> [![Travis Build
-Status](https://travis-ci.org/kamapu/taxlist.svg?branch=master)](https://travis-ci.org/kamapu/taxlist)
-[![Codecov test
-coverage](https://codecov.io/gh/kamapu/taxlist/branch/master/graph/badge.svg)](https://codecov.io/gh/kamapu/taxlist?branch=master)
-<br>
-[![CRAN\_downloads](http://cranlogs.r-pkg.org/badges/taxlist)](https://cran.r-project.org/package=taxlist)
-[![total
-downloads](http://cranlogs.r-pkg.org/badges/grand-total/taxlist)](https://cran.r-project.org/package=taxlist)
+The aim of `taxlist` is to provide an object structure for taxonomic lists and methods to display and handle the contained information. We welcome any interest to implement it or to contribute in the development of `taxlist`.
 
-The aim of `taxlist` is to provide an object structure for taxonomic
-lists and methods to display and handle the contained information. We
-welcome any interest to implement it or to contribute in the development
-of `taxlist`.
+Objects and functions of `taxlist` are also implemented as part of objects containing information from vegetation-plot databases (look at [vegtable](https://github.com/kamapu/vegtable)). The structure of `taxlist` objects is strongly inspired on data handled by [Turboveg](https://www.synbiosys.alterra.nl/turboveg) and on relational database models.
 
-Objects and functions of `taxlist` are also implemented as part of
-objects containing information from vegetation-plot databases (look at
-[vegtable](https://github.com/kamapu/vegtable)). The structure of
-`taxlist` objects is strongly inspired on data handled by
-[Turboveg](https://www.synbiosys.alterra.nl/turboveg) and on relational
-database models.
+![](README-figures/taxlist_model.png)<br/> **Figure:** Relational model for taxlist objects (see [Alvarez & Luebert 2018](https://doi.org/10.3897/BDJ.6.e23635)).
 
[…]
