+++ README.Rmd
@@ -174,7 +174,7 @@ summary(Easplist, "Cyperus papyrus")
 ````
 
 You can than insert insert in your document inline codes, such as
-`r sprintf('\x60r print_name(Easplist, 206)\x60')`, which will insert
+`r '&grave;r print_name(Easplist, 206)&grave;'`, which will insert
 `r print_name(Easplist, 206)` in your document (note that the number is the ID
 of the taxon concept in `Easplist`).
 Fort a second mention of the same species, you can then use
