### 2020-01-10 9e9876d
Final solution using html code in inline verbatim chuncks
