+++ README.Rmd
@@ -169,17 +169,17 @@ chunk:
 `r ''````{r  echo=FALSE, message=FALSE, warning=FALSE}
 library(taxlist)
 data(Easplist)
-summary(Easplist, "Cyperus papyrus")
 ```
 ````
 
 You can than insert insert in your document inline codes, such as
-<code>'&grave;r print_name(Easplist, 206)&grave;'</code>, which will insert
+<code>&grave;r print_name(Easplist, 206)&grave;</code>, which will insert
 `r print_name(Easplist, 206)` in your document (note that the number is the ID
 of the taxon concept in `Easplist`).
 Fort a second mention of the same species, you can then use
-<code>`r print_name(Easplist, 206, second_mention=TRUE)`</code>, which will
-insert `r print_name(Easplist, 206, second_mention=TRUE)` in your text.
+<code>&grave;r print_name(Easplist, 206, second_mention=TRUE)&grave;</code>,
+which will insert `r print_name(Easplist, 206, second_mention=TRUE)` in your
+text.
 
 
 ## Acknowledgements
