### 2020-01-10 #233 @maelle
Thanks for the updates!

I am sorry, I've just noticed you do not use [`roxygen2`](https://roxygen2.r-lib.org/) for producing documentation, which is a [requirement before review](https://devguide.ropensci.org/building.html#documentation). I haven't done the conversion from Rd to roxygen2 myself but you might find the [rd2roxygen package]( https://yihui.org/rd2roxygen/) useful. When using roxygen2 you won't have to updated NAMESPACE by hand. See e.g. the [chapter about man/ from the R Packages book](https://r-pkgs.org/man.html) and [the chapter about NAMESPACE](https://r-pkgs.org/namespace.html). I am really sorry for not seeing that earlier, but that change will be for the best.

Reg your questions

1. Not exactly sure, you could save one RDA and in the test compare the new RDA you produce to the existing one?
2. I think you need to use [`expect_output()`](https://testthat.r-lib.org/reference/expect_output.html)
