### 2020-01-10 #233 @maelle
I'd even recommend using [Markdown formatting for roxygen2](https://roxygen2.r-lib.org/articles/rd-formatting.html#turning-on-markdown-support) so you might need to run [roxygen2md](https://roxygen2md.r-lib.org/) after Rd2roxygen. :-) (that part is not a requirement, I just think it's easier to use Markdown formatting later, but you can disagree).
