### 2020-01-10 #233 @kamapu
Thank you, for your prompt answer. I just liked to announce you about the last version uploaded to master, to release me from the work, but I see, it is a never ending story...

I may have overseen the implementation of `roxygen` in the source. I have read about it but never given a try, thus again, something new for me and more time to get there...

Regarding the test on writing function, were will be the file get written? in the active working directory, meaning the main folder of the package source? In a "virtual" (temporary) working directory?
