### 2020-01-10 #233 @maelle
You won't regret learning about roxygen2, for this package and future packages. :wink: It's fine if it takes time, I can imagine you're busy.

The reference file would be in the test/ folder, and the one created during test in a temporary directory. You can use the same code [as in this helper file](https://github.com/lockedata/starters/blob/master/tests/testthat/helper.R) that'll create a temporary directory and delete it once the test is run. the `check=TRUE` argument of `tempdir()` was introduced in R 3.5.0.
