### 2020-01-10 bad2aa2
Test for `summary()`
