### 2020-01-18 f2c3635
Small change in vignettes
