+++ R/taxon_names.R
@@ -1,26 +1,187 @@
-# TODO:   Specific methods for slot taxonNames
-# 
-# Author: Miguel Alvarez
-################################################################################
-
+#' @name taxon_names
+#' @aliases taxon_names,taxlist-method
+#' @aliases taxon_names<-
+#' @aliases taxon_names<-,taxlist,data.frame-method
+#' @aliases add_synonym
+#' @aliases add_synonym,taxlist-method
+#' @aliases update_name
+#' @aliases update_name,taxlist,numeric-method
+#' @aliases delete_name
+#' @aliases delete_name,taxlist,numeric-method
+#' 
+#' @title Handle information on taxon usage names.
+#' 
+#' @description 
+#' The slot `taxonNames` in \code{\linkS4class{taxlist}} objects contains
+#' taxon usage names for the respective taxon.
+#' These functions assist on the access and modification of entries for names.
+#' 
+#' @param taxlist A \code{\linkS4class{taxlist}} object to be modified.
+#' @param value A data frame used as new slot `taxonNames` in `taxlist`.
+#' @param ConceptID Numeric vector indicating the concept ID to which the
+#'     synonyms will be added.
+#' @param TaxonName,AuthorName Character values used for the new names
+#'     (synonyms).
+#' @param UsageID Numeric vector indicating the taxon usage IDs to be updated.
+#' @param ... Further arguments passed among methods. In `update_name` are
+#'     vectors including the variables to be updated for the respective taxon
+#'     usage ID.
+#' 
+#' @details 
+#' The replacement method `taxon_names<-` is a quick alternative to include
+#' names in empty \code{\linkS4class{taxlist}} objects.
+#' 
+#' The function \code{'add_synonym'} works only for adding names to existing
+#' taxon concepts.
+#' For adding new taxon concepts as well you should use
+#' \code{\link{add_concept}}.
+#' 
+#' @return A data frame or, in the case of the replacement method, a
+#' \code{\linkS4class{taxlist}} object with modified slot \code{'taxonNames'}.
+#' 
+#' @author Miguel Alvarez \email{kamapu78@@gmail.com}
+#' 
+#' @seealso \code{\linkS4class{taxlist}}.
+#' 
+#' @examples 
+#' ## Display of slot 'taxonNames'
+#' Euclea <- subset(Easplist, charmatch("Euclea", TaxonName), slot="names",
+#'     keep_children=TRUE)
+#' summary(Euclea)
+#' taxon_names(Euclea)
+#' 
+#' ## Insert a synonym to Diospyros scabra
+#' summary(Easplist, "Diospyros scabra")
+#' Easplist <- add_synonym(Easplist, 51793, TaxonName="Maba scabra",
+#'     AuthorName="Chiov.")
+#' summary(Easplist, "Diospyros scabra")
+#' 
+#' ## Delete a synonym of Launaea cornuta
+#' summary(Easplist, "Launaea cornuta")
+#' Easplist <- delete_name(Easplist, 53821)
+#' summary(Easplist, "Launaea cornuta")
+#' 
+#' @rdname taxon_names
+#' @export 
 setGeneric("taxon_names",
         function(taxlist, ...)
             standardGeneric("taxon_names")
 )
 
-# Set method for taxlist
+#' @rdname taxon_names
+#' 
+#' @export 
 setMethod("taxon_names", signature(taxlist="taxlist"),
         function(taxlist, ...) taxlist@taxonNames
 )
 
-# Replacement methods
+#' @rdname taxon_names
+#' 
[…]
