+++ R/count_taxa.R
@@ -1,15 +1,49 @@
-# TODO:   Counting taxa by their names
-# 
-# Author: Miguel Alvarez
-################################################################################
-
-# Generic function
+#' @name count_taxa
+#' @aliases count_taxa,character-method
+#' @aliases count_taxa,factor-method
+#' @aliases count_taxa,taxlist-method
+#' 
+#' @title Count taxa within a taxlist object.
+#' 
+#' @description 
+#' Counting number of taxa within \code{\linkS4class{taxlist}} objects or
+#' character vectors containing taxon names.
+#' 
+#' @param object An object of class \code{\linkS4class{taxlist}}.
+#' @param rm.na Logical value, whether NAs have to be removed from the input
+#'     vector or not.
+#' @param level Character value indicating the taxonomic rank of counted taxa.
+#' @param ... further arguments passed among methods.
+#' 
+#' @details 
+#' This function is written by convenience in order to reduce code for counting
+#' taxa within \code{\linkS4class{taxlist}} objects and it is just a wrapper of
+#' \code{\link{length}}.
+#' 
+#' @return An integer with the number of taxa.
+#' 
+#' @author Miguel Alvarez.
+#' 
+#' @examples 
+#' library(taxlist)
+#' 
+#' ## factor method
+#' count_taxa(iris$Species)
+#' 
+#' ## taxlist method
+#' count_taxa(Easplist)
+#' count_taxa(Easplist, level="species")
+#' 
+#' @rdname count_taxa
+#' @export 
 setGeneric("count_taxa",
 		function(object, ...)
 			standardGeneric("count_taxa")
 )
 
-# Method for character vectors
+#' @rdname count_taxa
+#' 
+#' @export 
 setMethod("count_taxa", signature(object="character"),
 		function(object, rm.na=TRUE, ...) {
 			if(rm.na) object <- object[!is.na(object)]
@@ -17,7 +51,9 @@ setMethod("count_taxa", signature(object="character"),
 		}
 )
 
-# Method for factors
+#' @rdname count_taxa
+#' 
+#' @export 
 setMethod("count_taxa", signature(object="factor"),
 		function(object, rm.na=TRUE, ...) {
 			if(rm.na) object <- object[!is.na(object)]
@@ -25,7 +61,9 @@ setMethod("count_taxa", signature(object="factor"),
 		}
 )
 
-# Method for taxlist objects
+#' @rdname count_taxa
+#' 
+#' @export 
 setMethod("count_taxa", signature(object="taxlist"),
 		function(object, level, ...) {
 			if(missing(level))

[…]
