+++ tests/testthat/test-load_last.R
@@ -1,8 +1,3 @@
-# TODO:   Add comment
-# 
-# Author: Miguel Alvarez
-################################################################################
-
 context("loading a backup")
 
 test_that("load_last is working", {
