+++ tests/testthat/test-summary.R
@@ -1,8 +1,3 @@
-# TODO:   Add comment
-# 
-# Author: Miguel Alvarez
-################################################################################
-
 context("Testing summary outputs")
 
 test_that("Summaries are produced", {
@@ -11,4 +6,6 @@ test_that("Summaries are produced", {
 			result <- evaluate_promise(summary(Easplist, "Cyperus pap"),
 					print=TRUE)
 			expect_equal(grepl("Cyperus", result$output), TRUE)
+			expect_error(summary(Easplist, 1000000000))
+			expect_error(summary(Easplist, 1, display="artist"))
 		})
