+++ R/clean.R
@@ -23,13 +23,12 @@
 #' @author Miguel Alvarez.
 #' 
 #' @examples 
-#' library(taxlist)
-#' data(Easplist)
-#' 	
 #' ## Direct manipulation of slot taxonRelations generates an invalid object
 #' Easplist@taxonRelations <- Easplist@taxonRelations[1:5,]
+#' \dontrun{
 #' summary(Easplist)
-#' 	
+#' }	
+#' 
 #' ## Now apply cleaning
 #' Easplist <- clean(Easplist)
 #' summary(Easplist)
