+++ tests/testthat.R
@@ -1,8 +1,3 @@
-# TODO:   Script for test suit
-# 
-# Author: Miguel Alvarez
-################################################################################
-
 # browseURL("https://kbroman.org/pkg_primer/pages/tests.html")
 # browseURL(file.path("https://walczak.org/2017/06",
 # 		"how-to-add-code-coverage-codecov-to-your-r-package/"))
