### 2020-01-26 a6ae718
A statistic section added to README
