+++ tests/testthat/test-count_taxa.R
@@ -3,9 +3,9 @@ context("counting taxa")
 test_that("count_taxa is working", {
 			expect_equal(count_taxa(iris$Species), length(unique(iris$Species)))
 			expect_equal(count_taxa(Easplist), nrow(Easplist@taxonRelations))
-			expect_equal(count_taxa(Easplist, "species"),
+			expect_equal(count_taxa(Easplist, level="species"),
 					with(Easplist@taxonRelations,
 							length(Level[paste(Level) == "species"])))
-			expect_error(count_taxa(Easplist, "superspecies"))
+			expect_error(count_taxa(Easplist, level="superspecies"))
 		}
 )
