+++ R/imports.R
@@ -1,6 +1,6 @@
 #' @importFrom  methods new slot slotNames validObject
 #' @importFrom utils object.size
-#' @importFrom stats aggregate na.omit sd
+#' @importFrom stats aggregate na.omit sd as.formula terms
 #' @importFrom vegdata tv.home
 #' @importFrom taxize tnrs
 #' @importFrom foreign read.dbf
