+++ R/count_taxa.R
@@ -1,6 +1,6 @@
 #' @name count_taxa
-#' @aliases count_taxa,character-method count_taxa,factor-method
-#'     count_taxa,taxlist-method
+#' @aliases count_taxa,character,missing-method count_taxa,factor,missing-method
+#'     count_taxa,taxlist,missing-method count_taxa,formula,taxlist-method
 #' 
 #' @title Count taxa within a taxlist object.
 #' 
@@ -8,7 +8,9 @@
 #' Counting number of taxa within \code{\linkS4class{taxlist}} objects or
 #' character vectors containing taxon names.
 #' 
-#' @param object An object of class \code{\linkS4class{taxlist}}.
+#' @param object An object containing a taxonomic list or a formula.
+#' @param data An object of class \code{\linkS4class{taxlist}} in the `formula`
+#'     method.
 #' @param rm.na Logical value, whether NAs have to be removed from the input
 #'     vector or not.
 #' @param level Character value indicating the taxonomic rank of counted taxa.
@@ -33,18 +35,21 @@
 #' count_taxa(Easplist)
 #' count_taxa(Easplist, level="species")
 #' 
+#' ## using a formula
+#' count_taxa(~ lf_behn_2018, Easplist)
+#' 
 #' @rdname count_taxa
 #' @export 
 setGeneric("count_taxa",
-		function(object, ...)
+		function(object, data, ...)
 			standardGeneric("count_taxa")
 )
 
 #' @rdname count_taxa
 #' 
 #' @export 
-setMethod("count_taxa", signature(object="character"),
-		function(object, rm.na=TRUE, ...) {
+setMethod("count_taxa", signature(object="character", data="missing"),
+		function(object, data, rm.na=TRUE, ...) {
 			if(rm.na) object <- object[!is.na(object)]
 			return(length(unique(object)))
 		}
@@ -53,8 +58,8 @@ setMethod("count_taxa", signature(object="character"),
 #' @rdname count_taxa
 #' 
 #' @export 
-setMethod("count_taxa", signature(object="factor"),
-		function(object, rm.na=TRUE, ...) {
+setMethod("count_taxa", signature(object="factor", data="missing"),
+		function(object, data, rm.na=TRUE, ...) {
 			if(rm.na) object <- object[!is.na(object)]
 			return(count_taxa(paste(object)))
 		}
@@ -63,8 +68,8 @@ setMethod("count_taxa", signature(object="factor"),
 #' @rdname count_taxa
 #' 
 #' @export 
-setMethod("count_taxa", signature(object="taxlist"),
-		function(object, level, ...) {
+setMethod("count_taxa", signature(object="taxlist", data="missing"),
+		function(object, data, level, ...) {
 			if(missing(level))
 				n_taxa <- nrow(object@taxonRelations) else {
 				if(!level %in% levels(object))
@@ -74,5 +79,57 @@ setMethod("count_taxa", signature(object="taxlist"),
 								paste(object@taxonRelations$Level) == level,])
 			}
 			return(n_taxa)
+		}
+)
+
+#' @rdname count_taxa
+#' 
+#' @param include_na Logical value indicating whether `NA` values in a taxon
+#'     trait should be considered for counting taxa or just ignored (only
+#'     used in `formula` method).
+#' @param suffix Character value used as suffix for the counted rank in the
+#'     output data frame (only used in `formula` method).
+#' 
+#' @export 
+setMethod("count_taxa", signature(object="formula", data="taxlist"),
+		function(object, data, include_na=FALSE, suffix="_count", ...) {
+			# Some checks
+			nr_response <- attr(terms(object), "response")
+			if(nr_response > 1)
[…]
