+++ README.Rmd
@@ -11,7 +11,7 @@ output:
 knitr::opts_chunk$set(
   collapse=TRUE,
   comment="#>",
-  fig.path="README-"
+  fig.path="README-figures/"
 )
 ```
 
@@ -182,12 +182,27 @@ which will insert `r print_name(Easplist, 206, second_mention=TRUE)` in your
 text.
 
 
-## Statistical Assessment
+## Descriptive Statistics
 
 Information located in the slot **taxonTraits** are suitable for statistical
 assessments.
+For instance, in the installed object `Easplist` a column called
+**lf_behn_2018** includes a classification of macrophytes into different life
+forms.
+To know the frequency of these life forms in the `Easplist`, we can use the
+function `count_taxa()`:
+
+```{r}
+# how man taxa in 'Easplist'
+count_taxa(Easplist)
+
+# frequency of life forms
+count_taxa(~ lf_behn_2018, Easplist)
+```
+
 Furthermore, taxonomic information can be also transferred to this slot using
 the function `tax2traits()`.
+By this way we will make taxonomic ranks suitable for frequency calculations.
 
 ```{r}
 Easplist <- tax2traits(Easplist, get_names=TRUE)
@@ -197,28 +212,22 @@ head(Easplist@taxonTraits)
 Note that the respective parental ranks are inserted in the table
 **taxonTraits**, which contains the attributes of the taxa.
 In the two next command lines, we will produce a subset with only members of
-the family Cyperaceae and then extract only species.
+the family Cyperaceae and then calculate the frequency of species per genera.
 
 ```{r}
 Cype <- subset(Easplist, family == "Cyperaceae", slot="taxonTraits")
-Cype <- subset(Cype, Level == "species", slot="taxonRelations")
+Cype_stat <- count_taxa(species ~ genus, Cype)
 ```
 
-Now, we can count the number of species in each genus of the family Cyperaceae
-and display them in a bar plot:
+Now, we can sort them to produce a nice bar plot.
 
-```{r}
-Cype_stat <- summary(as.factor(Cype@taxonTraits$genus))
-Cype_stat <- Cype_stat[order(Cype_stat, decreasing=TRUE)]
+```{r genera_bar}
+Cype_stat <- Cype_stat[order(Cype_stat$species_count, decreasing=TRUE),]
 
 par(las=2, mar=c(10,3,1,1))
-barplot(Cype_stat)
+with(Cype_stat, barplot(species_count, names.arg=genus))
 ```
 
-Such kind of calculations can be also implemented in functions once working
-with biodiversity records, for instance in the functions `trait_stats()` and
-`trait_proportion()` of the package
-`[vegtable](https://github.com/kamapu/vegtable)`.
 
 ## Acknowledgements
 
