+++ NAMESPACE
@@ -94,8 +94,10 @@ importFrom(methods,slot)
 importFrom(methods,slotNames)
 importFrom(methods,validObject)
 importFrom(stats,aggregate)
+importFrom(stats,as.formula)
 importFrom(stats,na.omit)
 importFrom(stats,sd)
+importFrom(stats,terms)
 importFrom(stringdist,stringsim)
 importFrom(taxa,taxon_ids)
 importFrom(taxa,taxon_ranks)

[…]
