### 2020-02-02 #233 @kamapu
Dear @maelle
I managed to "roxygenize" the package. You are right, it makes a lot of sense to implement it, though it has been also very helpful to have the experience doing documentation by hand.
What is next? :)
