### 2020-02-02 #233 @maelle
Awesome, congrats! I too first learned to write docs by hand but am not sure I wouldn't have preferred to be shown roxygen2 first 😁

A remaining point was "Insert comparison with taxa in README", have you made progress on that?
