### 2020-02-03 #233 @kamapu
I tried but using a different approach then the one suggested:
1) The first paragraph in the chapter **Similar Packages** summarizes the differences between `taxlist` and `taxa` with a link to the detailed discussion.
2) I also inserted two chapters called **Rmarkdown Integration** and **Descriptive Statistics** that do not directly mention `taxa` but attempt to highlight special features of `taxlist`, which may not be considered in any other package dealing with taxonomic information.

Pleas, let me know if this is OK so or if some additions are required.

I also remind you, that the functions written to transform `taxlist` objects into `taxmap` and vice versa were removed from the master branch since they are not working with the current CRAN version of `taxa`.
I will reinsert those functions once this package is accordingly updated.
