### 2020-02-03 #233 @maelle
Thank you. I'm not sure it's sufficient to answer the question "how would an R user with a taxonomy problem at hand know how to choose one of the two packages?", especially as it'd demand a potential user to read and digest information from a GitHub issue thread. We were especially hoping both packages would have the same summary/info in their README. @sckott @zachary-foster could you please comment on that + on when taxa will be updated on CRAN? Thanks all.

Furthermore, @kamapu, to help my understanding and my reviewer search when time comes, feel free to point me to use cases by other people. :-)
