### 2020-02-04 #233 @zachary-foster
> "how would an R user with a taxonomy problem at hand know how to choose one of the two packages?" ... We were especially hoping both packages would have the same summary/info in their README.

I am not really sure to be honest. Its been a bit since I tried out `taxlist`. I might have to revisit it to answer that. @kamapu can correct me if I am wrong, but perhaps the main difference is that `taxa` is targeted towards developers that want to make an R package that uses taxonomic data, but does not want to make classes and manipulation functions from scratch. `taxa` is like `tibble` + `dplyr` for taxonomic data. Non-developer users would interact with the classes and functions defined by `taxa`, but would do so in the context of another package, much like tibbles are used in packages besides `tibble` and users of tibbles might not even know which package defines them.

> on when taxa will be updated on CRAN?

I will try to get an update out this week
