### 2020-02-08 3896e0d
Integration of rmarkdown in documentation
