### 2020-02-26 kamapu/taxlist#2 @zachary-foster
New version of `taxa` is on CRAN now. Sorry for the delay!
