### 2020-04-10 82958de
New argument in function 'backup_object()'
