+++ DESCRIPTION
@@ -74,4 +74,4 @@ Collate:
     'taxlist-package.R'
     'StartMessage.R'
 VignetteBuilder: knitr
-RoxygenNote: 7.0.2
+RoxygenNote: 7.1.0
