### 2020-04-11 0ce1770
small changes in internal function
