### 2020-04-11 c88a500
Internal function 'sort_backups' written for general use
