### 2020-04-19 e7f10d1
New function 'taxlist2taxmap()'
