+++ R/taxlist2taxmap.R
@@ -1,0 +1,173 @@
+#' @name taxlist2taxmap
+#' @rdname taxlist2taxmap
+#' 
+#' @title Conversion among taxlist and taxmap objects
+#' 
+#' @description 
+#' Exchange of data between the packages `taxlist` and `taxa`.
+#' 
+#' This function should be used for the exchange of data between the packages
+#' [taxlist-package] and `taxa`.
+#' 
+#' @param taxlist Input object of class [taxlist-class].
+#' @param taxmap Input object of class [taxmap].
+#' @param relations,traits,synonyms,views Character values indicating the names
+#'     of data frames in the [taxmap] object at `data`, which should be
+#'     used for the slots taxonRelations, taxonTraits, taxonNames, and
+#'     taxonViews, respectively.
+#' @param reindex Logical value indicating whether taxon IDs should be assigned
+#' 	   anew or not.
+#' @param ... Additional arguments passed among methods.
+#' 
+#' @return 
+#' Depending on the applied function, either a [taxlist-class] or a `Taxmap`
+#' object.
+#' 
+#' @author 
+#' Miguel Alvarez (\email{kamapu78@@gmail.com}) and Zachary Foster
+#' (\email{zacharyfoster1989@@gmail.com}).
+#' 
+#' @examples 
+#' library(taxa)
+#' 
+#' @exportMethod taxlist2taxmap
+#' 
+setGeneric("taxlist2taxmap",
+		function(taxlist, ...)
+			standardGeneric("taxlist2taxmap")
+)
+
+#' @rdname taxlist2taxmap
+#' 
+#' @aliases taxlist2taxmap,taxlist-method
+#' 
+setMethod("taxlist2taxmap", signature(taxlist="taxlist"),
+		function(taxlist, ...) {
+			# Use the edge list to start making the intput object
+			taxlist_df <- accepted_name(taxlist)
+			taxlist_df$Parent <- with(taxlist@taxonRelations,
+					Parent[match(taxlist_df$TaxonConceptID, TaxonConceptID)])
+			taxlist_df$Level <- as.character(taxlist_df$Level)
+			obj <- taxa::parse_edge_list(taxlist_df, taxon_id="TaxonConceptID",
+					supertaxon_id="Parent", taxon_name="TaxonName",
+					taxon_rank="Level")
+			names(obj$data) <- c("relations")
+			obj$data$relations <- with(obj$data,
+					relations[,colnames(relations) != "TaxonConceptID"])
+			# Set taxon authorities
+			obj$set_taxon_auths(with(taxlist_df,
+							AuthorName[match(obj$taxon_ids(), TaxonConceptID)]))
+			# Add traits table to the taxlistect
+			obj$data$traits <- taxlist@taxonTraits
+			names(obj$data$traits)[1] <- "taxon_id"
+			obj$data$traits$taxon_id <- paste(obj$data$traits$taxon_id)
+			# Add views table to the taxlistect
+			obj$data$views <- taxlist@taxonViews
+			# Add synonyms in a table
+			obj$data$synonyms <- synonyms(taxlist)[,c("TaxonConceptID",
+							"TaxonUsageID", "TaxonName", "AuthorName")]
+			names(obj$data$synonyms) <- c("taxon_id", "TaxonUsageID", "synonym",
+					"synonym_authority")
+			obj$data$synonyms$taxon_id <- paste(obj$data$synonyms$taxon_id)
+			
+			return(obj)
+		}
+)
+
[…]
