+++ data-raw/0_test.R
@@ -1,0 +1,7 @@
+# TODO:   Working script for testing the package 'taxlist'
+# 
+# Author: Miguel Alvarez
+################################################################################
+
+
+
