+++ man/taxlist-package.Rd
@@ -1,8 +1,8 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/taxlist-package.R
 \docType{package}
-\name{taxlist}
-\alias{taxlist}
+\name{taxlist-package}
+\alias{taxlist-package}
 \title{taxlist: Handling taxonomic lists.}
 \description{
 The class \linkS4class{taxlist} is defined in this package using the
