+++ NAMESPACE
@@ -12,6 +12,7 @@ export(replace_idx)
 export(replace_na)
 export(replace_view)
 export(replace_x)
+export(taxmap2taxlist)
 export(tv2taxlist)
 exportClasses(taxlist)
 exportMethods("$")
@@ -45,6 +46,7 @@ exportMethods(subset)
 exportMethods(summary)
 exportMethods(synonyms)
 exportMethods(tax2traits)
+exportMethods(taxlist2taxmap)
 exportMethods(taxon_names)
 exportMethods(taxon_relations)
 exportMethods(taxon_traits)
