### 2020-04-19 e22d07e
New version for ROpenSci
