+++ R/replace_x.R
@@ -37,7 +37,8 @@
 #' replace_x(letters, c("b", "p", "f"), c("bee", "pork", "fungus"))
 #' 
 #' ## Replace values using indices
-#' replace_idx(letters, 1:length(letters), c(2,7,17), c("second", "seven", "seventeenth"))
+#' replace_idx(letters, 1:length(letters), c(2,7,17), c("second", "seven",
+#'     "seventeenth"))
 #' 
 #' ## Replace values if they are NAs
 #' letters[2] <- NA
@@ -46,7 +47,8 @@
 #' ## Merge data frames including new columns
 #' data(iris)
 #' iris$Species <- paste(iris$Species)
-#' new_iris <- data.frame(Species=rep("humilis", 2), Height=c(15,20), stringsAsFactors=FALSE)
+#' new_iris <- data.frame(Species=rep("humilis", 2), Height=c(15,20),
+#'     stringsAsFactors=FALSE)
 #' insert_rows(iris, new_iris)
 #' 
 #' @rdname replace_x
