+++ cran-comments.md
@@ -1,0 +1,6 @@
+## Test environments
+* ubuntu 12.04 (on travis-ci), R 3.1.2
+* win-builder (devel and release)
+
+## R CMD check results
+There were no ERRORs, WARNINGs or NOTEs. 

[…]
