### 2020-04-20 381bdd5
Deprecated functions properly documented
