+++ README.md
@@ -21,22 +21,23 @@ downloads](http://cranlogs.r-pkg.org/badges/grand-total/taxlist)](https://cran.r
 The aim of `taxlist` is to provide an object structure for taxonomic
 lists, as well as functions for handling and displaying the information
 contained in such objects. The structure of `taxlist` objects is
-inspired on data handled by
-[Turboveg](https://www.synbiosys.alterra.nl/turboveg) and on relational
-database models.
+inspired on the structure of data handled by
+[Turboveg](https://www.synbiosys.alterra.nl/turboveg) and relational
+databases.
 
 ![](README-figures/taxlist_model.png)<br/> **Figure:** Relational model
 for `taxlist` objects (see [Alvarez & Luebert
 2018](https://doi.org/10.3897/BDJ.6.e23635)).
 
-The functions in this package has being used to structure and clean data
-stored in [SWEA-Dataveg](https://kamapu.github.io/sweadataveg.html), a
+The functions in this package has being used to structure and
+cross-check consistency of data stored in
+[SWEA-Dataveg](https://kamapu.github.io/sweadataveg.html), a
 vegetation-plot database for Eastern Africa.
 
 ## Installing taxlist
 
 This package is available from the Comprehensive R Archive Network
-(**CRAN**) and can be directly installed in an R-session:
+(**CRAN**) and can be directly installed within an R-session:
 
 ``` r
 install.packages("taxlist", dependencies=TRUE)
@@ -53,10 +54,10 @@ install_github("kamapu/taxlist", build_vignette=TRUE)
 
 ## Building taxlist Objects
 
-Objects can be built step-by-step as in this example taking as reference
-the “Ferns of Chile” (original in Spanish: “Helechos de Chile”) by
-**Gunkel (1984)**. The first step will be to generate an empty `taxlist`
-object:
+Objects can be built step-by-step as in the following example. For it,
+we will use as reference the “Ferns of Chile” (original in Spanish:
+“Helechos de Chile”) by **Gunkel (1984)**. We will create an empty
+`taxlist` object using the function `new()`:
 
 ``` r
 library(taxlist)
@@ -73,15 +74,16 @@ summary(Fern)
 #> taxon views: 0
 ```
 
-We start defining the required taxonomic ranks. In such case, the levels
-have to be provided from the lowest to highest hierarchical level:
+Then we have to set the respective taxonomic ranks. In such case, the
+levels have to be provided from the lowest to highest hierarchical
+level:
 
 ``` r
 levels(Fern) <- c("variety","species","genus")
 ```
 
 For convenience, we start inserting taxa with their respective names in
-a top-down direction. We use then the function `add_concept()` to add a
+a top-down direction. We will use the function `add_concept()` to add a
 new taxon. Note that the arguments `TaxonName`, `AuthorName`, and
 `Level` are used to provide the name of the taxon, the authority of the
 name and the taxonomic rank, respectively.
@@ -132,10 +134,11 @@ Fern <- add_concept(Fern,
 ```
 
 You may have realized that the function `summary()` is applied to
-provide on the one side a display of meta-information for the `taxlist`
-object, and on the other side to show a detail of the taxa included in
-the object. In the later case adding the keyword `"all"` in the second
-argument will show all the handled taxa.
+provide on the one side a display of meta-information for the whole
+`taxlist` object, and on the other side to show a detail of the taxa
+included in the object. In the later case adding the keyword `"all"` as
+second argument, the summary will show a detailed information for every
+taxon included in the object.
 
 ``` r
 summary(Fern)
@@ -194,33 +197,33 @@ summary(Fern, "all")
 
 ## Similar Packages
 
-The package `taxlist` shares similar objectives than the package
[…]
