+++ man/replace_view.Rd
@@ -1,11 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/deprecated-functions.R
-\name{replace_view}
-\alias{replace_view}
-\title{Replace taxon views in concepts.}
-\usage{
-replace_view()
-}
-\description{
-Replace taxon views in concepts.
-}
