+++ man/add_level.Rd
@@ -1,11 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/deprecated-functions.R
-\name{add_level}
-\alias{add_level}
-\title{Add level information.}
-\usage{
-add_level()
-}
-\description{
-Add level information.
-}

[…]
