### 2020-04-20 #233 @kamapu
@maelle I managed to produce a new version of `taxlist`, which may consider all tasks requested for its submission to **ROpenSci**.

- [x] Package documentation written with `roxygen2`.
- [x] Test suite implemented with `covr` and coverage icreased to over 90%.
- [x] Functions `taxlist2taxmap()` and `taxmap2taxlist()` developed for the exchange of objects with `taxa`.
- [x] A list of hints guiding users for the choice of `taxlist` vs. `taxa`.

On the later case, I wrote an itemized list in [README](https://github.com/kamapu/taxlist), chapter **Similar Packages**.
Note that since I am not really working much with `taxa`, I'm not able to provide a neutral comparison between the two packages. The same is valid for the developers of `taxa` in the other direction. Thus I rather provide examples of applications, where the users may prefer or have to use `taxlist` instead of `taxa`.
