### 2020-04-25 #233 @maelle
You can still do changes, just don't decrease the code coverage now that it's so good 😉 We'll update threads when we are back to normal operation.
