### 2020-05-25 #233 @kamapu
Welcome back! This are good news and a signal that a piece of life is returning to normality.
I have some problems with CRAN because the encoding of example data (Encoding is becoming my worse nightmare). I may solve it by the end of the week and produce a new release.
