### 2020-05-25 #233 @maelle
Ok, I'll wait until then! You can ask questions on [rOpenSci forum](https://discuss.ropensci.org/) if needed. Good luck! (and yeah encoding is hard!)
