### 2020-05-28 7749572
Solving problems with Encoding in data set Easplist
