### 2020-05-29 1fa5c8c
Adjusting validity checks
