### 2020-05-30 3b2269d
check done in Latin-1 environment
