### 2020-05-30 #233 @kamapu
The issue with encoding was solved in the last commit, @maelle Now we can proceed.
