### 2020-06-01 #233 @maelle
Thanks!

Is https://github.com/kamapu/taxlist/blob/master/TODO.md out-of-date? If so could you please remove it to not distract reviewers?
