### 2020-06-01 #233 @maelle
Reviewers assigned! Thanks a lot @mcsiple & @levisc8 for agreeing to review. Your reviews are due on 2020-06-22.

As a reminder here are links to the [reviewer guide](https://devguide.ropensci.org/reviewerguide.html) and [review template](https://devguide.ropensci.org/reviewtemplate.html).
