### 2020-06-12 #233 @maelle
Thanks a ton for your thorough review @levisc8! :rocket: :pray:

A few notes from me

* No need to set up a pkgdown website now since there'll be one built for the package once it is approved. Unless @kamapu wants feedback on the navbar and reference index grouping for instance. [We're however thinking about this](https://github.com/ropensci/dev_guide/issues/261), I'm not saying it's a bad comment, on the contrary. :wink:

* Regarding the paths in tests, in case it is relevant (sorry if I'm missing subtleties). In rtimicropem I put such data in inst/exdata and [access it from tests](https://github.com/ropensci/rtimicropem/blob/86475e13cbbbb974ca336b8d84daa45e66e2a410/tests/testthat/test-convert_output.R#L9).

* Just adding the link to [the dev guide section about CONTRIBUTING.md](https://devguide.ropensci.org/collaboration.html#contributing-guide)
