### 2020-06-19 #233 @kamapu
Wow! I was away of GitHub for a while (also thanks the computer in my home-office that I had to install many times). Thanks @levisc8 for the review. I assume, I can start answering and implementing suggestions, isn't it @maelle ?
