### 2020-06-22 #233 @mcsiple
Hi @maelle, @kamapu, and @zachary-foster, here's my review! Please let me know if you have any questions about anything I wrote.

I did not find any Contributing Guidelines in the README or DESCRIPTION

>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](https://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](https://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

###### Local install
Local install worked fine; I had some classic issues with updating {rlang} and {ps} and a couple of other packages but once I had all the dependencies there weren't any problems installing locally. (I say "classic" issues because I have often had problems when {rlang} needs to be updated. Not sure why this is.)

###### Install from GitHub
Install from GitHub works smoothly.

###### Good practices checklist
`r goodpractice::gp()` returns three good practices that are missing from {taxlist}:
- fix this R CMD check NOTE: Note: found 137 marked UTF-8 strings
- fix this R CMD check WARNING: LaTeX errors when creating PDF version. This typically indicates Rd problems.
- fix this R CMD check ERROR: Re-running with no redirection of stdout/stderr. Hmm ... looks like a package Error in texi2dvi(file = file, pdf = TRUE, clean = clean, quiet = quiet, : pdflatex is not available Error in texi2dvi(file = file, pdf = TRUE, clean = clean, quiet = quiet, : pdflatex is not available Error in running `tools::texi2pdf()`

I am not sure what exactly is happening with any of these three but it sounds like an issue with PDF creation, which has come up before (see "Checks on package source").

###### Metadata files
- #### [README](https://github.com/kamapu/taxlist)
- #### [DESCRIPTION](https://github.com/kamapu/taxlist/blob/master/DESCRIPTION)
- #### [NAMESPACE](https://github.com/kamapu/taxlist/blob/master/NAMESPACE)

Some of the tests failed when I ran test(pkg_dir) (See "Package tests & checks" section)

---

#### Key review checks

- [X]  Does the code comply with **general principles in the [Mozilla reviewing guide](https://mozillascience.github.io/codeReview/review.html)**?
- [X]  Does the package **comply with the [ROpenSci packaging guide](https://ropensci.github.io/dev_guide/building.html)**?
- [X]  Are there **improvements** that could be made to the **code style?**
- [ ]  Is there **code duplication** in the package that should be reduced?
- [ ]  Are there **user interface improvements** that could be made?
- [ ]  Are there **performance improvements** that could be made?
- [X]  Is the [**documentation**](https://ropensci.github.io/dev_guide/building.html#documentation) (installation instructions/vignettes/examples/demos) **clear and sufficient**?
(Note: the )

Thank you for the opportunity to review this package. It seems to be used very widely already, so I know it is a useful tool. Like @levisc8 , I don't do a lot of taxonomy. I tend to work with existing taxonomic classifications for things (and use taxonomy more as a unique ID to get other information like stock status), but I tried to review the package as thoroughly and usefully as I could, but I urge the package authors to interpret my comments with that lens, and not to struggle to incorporate certain changes if you know they won't improve the functionality for everyday users of the package, or if it seems that I have missed the point on that topic.

#### Package checks & tests

##### Checks
0 errors and 1 warning returned from `r devtools::check()` : "`The warning is "'qpdf' is needed for checks on size reduction of PDFs`" (probably not important, but just so you have it!)

##### Tests
I ran into a couple errors and warnings in this section, which I see @levisc8 has suggested some useful solutions for. I think they're just coming from how the directories are set up for the checks and the examples.

`devtools::check(pkg_dir)` returns 4 failures and 2 warnings.

I got 4 failures:
* R:3 coerce objects to taxlist
* R:4 loading a backup ("requested backup missing from working directory")
	cannot open compressed file for examples.Rda
	It looks like examples.Rda is not where check() is looking for it.
	Note: I think this may be because the directory in test-taxon_relations.R should go to inst/taxlist_examples/examples.Rda. Currently that line is missing the inst/ part.
* R:5 unable to open DBF file in importing Turboveg datasets
	I'm not sure why this is happening. It looks like it's set up properly in the test (no directory issues). I saw one little post online that said sometimes Windows expects the .DBF to be capitalized in the file extension but changing it manually did not affect the test outcome for me. If I had more time I would test on a Mac too, and get to the bottom of this!

I got 3 warnings from `devtools::check()`:
* R:3 "no such file as cyperus/names.csv"
	For some reason `test()` is looking in taxlist/cyperus/ for names.csv instead of taxlist/inst/cyperus I think this directory is on line 3 of test-df2.taxlist.R. Like the R:4 error above, inst is missing from the directory.
* R:4 Another warning about the location of examples.Rda

##### Statement of need
My main question/confusing is that it was hard for me to figure out who the target user was, and how this user group overlaps (or doesn't) with users of {taxa}. I thought at first that taxlist might be a helper package for people who are building their own taxonomy packages, a replacement for {taxa}, or perhaps just provide a tool for classification in S4. The DESCRIPTION file suggests that it is specifically for users who are getting taxonomic data from Turboveg. Whatever you anticipate the dominant use for taxlist will be, that should be at the top of the README and the vignettes. I found the paper introducing the package to give a nice succinct [statement of need](https://search.proquest.com/docview/2282193678?pq-origsite=gscholar&fromopenview=true); that one could potentially just be tweaked a little and added to the README.

If people using this package will most likely be modifying/importing existing taxonomy lists (i.e., from Turboveg), I think there should be an example or two in the vignettes that show users how this is done. This seems to be a primary functionality.

##### Help files
I worked through the example in the vignette. Here are a few comments:
- It would be useful to have some information about how to create a taxlist that contains more than just species names. For example, if Cross is a dataframe with a "Genus" or a "Family" column (or both!), how might one build a taxlist from it? How can a user deal with missing values or mixed info (e.g., a Genus name accidentally included in the wrong column)? This may be especially relevant for old taxon lists like the one shown in the example.
- Is the information in `Cross[,2:8]` not incorporated anywhere in the result of `Cross[,"TaxonName"]`? If so, could someone theoretically use taxlist to create a brand new taxonomy with its own organizational structure? I think this is the case but it would be good to make it more obvious.
- I cannot tell if Codes is an object that exists by default in order to provide unique identifiers for each taxon, or whether it is related to a specific taxlist object.

##### Tiny edits I noticed as I was going along
None of these little copy edits will affect anything about the clarity of the documentation but I noticed them as I went along and thought I would document them anyway.
- Missing "n" in "in" on 1st line of Getting Started vignette
- Small correction: `taxon_relations()` is misspelled in the help file for taxon_relations (under 'Details')
- vignette section 3.4: 'information can get lost once `subset()` has been applied'

#### Session Info
```r
Platform: x86_64-w64-mingw32/x64 (64-bit)
Running under: Windows 10 x64 (build 18363)

[1] LC_COLLATE=English_United States.1252  LC_CTYPE=English_United States.1252
[3] LC_MONETARY=English_United States.1252 LC_NUMERIC=C
[5] LC_TIME=English_United States.1252

[1] devtools_2.3.0 usethis_1.6.1  magrittr_1.5

 [1] prettyunits_1.1.1 ps_1.3.3          fansi_0.4.1       rprojroot_1.3-2   withr_2.2.0
 [6] digest_0.6.25     crayon_1.3.4      assertthat_0.2.1  R6_2.4.1          backports_1.1.7
[11] rlang_0.4.6       cli_2.0.2         remotes_2.1.1     rstudioapi_0.11   fs_1.4.1
[16] testthat_2.3.2    callr_3.4.3       ellipsis_0.3.1    desc_1.2.0        tools_4.0.0
[21] glue_1.4.1        pkgload_1.1.0     processx_3.4.2    compiler_4.0.0    pkgbuild_1.0.8
[26] sessioninfo_1.1.1 memoise_1.1.0
```r
