### 2020-06-23 #233 @kamapu
OK, thank you to @mcsiple and @levisc8 for your valuable comments to `taxlist`. I guess, I should now start editing the package in order to answer to your questions. Isn't it, @maelle ?
