### 2020-06-24 #233 @maelle
@kamapu yes, indeed. You can continue the conversation here but at the end of your edits we expect a comment that responds to both reviews. (see the second to last point in https://devguide.ropensci.org/guide-for-authors.html)
