### 2020-07-02 #233 @kamapu
I hope, it is OK to save it here and edit while I'm answering, thus, ignore the text below the **TODO** tag:

# TODO

Response in edition...

# Response to levisc8

### Community guidelines

* [x]  Include contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`).

**Response:**

>
> > ##### For packages co-submitting to JOSS
> >
> > * [ ]  The package has an **obvious research application** according to [JOSS's definition](https://joss.theoj.org/about#submission_requirements)
> >
> > The package contains a `paper.md` matching [JOSS's requirements](https://joss.theoj.org/about#paper_structure) with:
> >
> > * [ ]  **A short summary** describing the high-level functionality of the software
> > * [ ]  **Authors:**  A list of authors with their affiliations
> > * [ ]  **A statement of need** clearly stating problems the software is designed to solve and its target audience.
> > * [ ]  **References:** with DOIs for all those that have one (e.g. papers, datasets, software).
>
> #### Functionality
>
>     * [x]  **Installation:** Installation succeeds as documented.
>
>     * [x]  **Functionality:** Any functional claims of the software been confirmed.
>
>     * [x]  **Performance:** Any performance claims of the software been confirmed.
>
>     * [ ]  **Automated tests:** Unit tests cover essential functions of the package
>       and a reasonable range of inputs and conditions. All tests pass on the local machine.
>
>     * [ ]  **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines
>
>
> #### Final approval (post-review)
>
>     * [ ]  **The author has responded to my review and made changes to my satisfaction. I recommend approving this package.**
>
>
> Estimated hours spent reviewing: 9
>
>     * [x]  Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer ("rev" role) in the package DESCRIPTION file.
>
>
> ### Review Comments
>
> The package authors have provided an excellent relational model for taxonomy. With 12k downloads and nearly 1k/month, it is already in use by the both author and quite a few others, and has already generated great value to the community. In the interest of full disclosure, I don't do much taxonomy work, so I may not be its target audience. Nonetheless, I found it pretty straightforward to use, and hope to have an opportunity to do so for my own work soon!
> #### Checks + Tests + goodpractice::gp()
>
> `devtools::check()` yields no errors, warnings, or notes, but the unit tests fail when run with `devtools::test()`. It appears to be an issue with the usage of `file.path(path.package(..., package = 'taxlist'))` calls in the unit tests. These return the file path to the source package in `test`, and the file path to the built package in `check`. In turn, I _think_ this represents the difference between `load_all()` used internally in `test` and `build` used internally in `check`.
>
> In every case except for `test-load_last.R`, you can replace the `file.path(path.package(...))` with `system.file('dir_within_pkg', 'file_name', package = 'taxlist')` and the tests will work both interactively and in `check`. This approach doesn't work for the `load_last` test because the function takes a directory path as opposed to a complete file path, and I'm not sure how to go about getting that without some nasty `gsub`ing, or adding another dependency like `fs`. One approach could be to generate a temporary directory, save an object in there, and then test `load_last` all within the test script itself. I'm not sure if this is best practice or not, and would request some guidance from others on that front.
>
> I'm happy to create a pull request with the `system.file()` replacements if you'd like!
>
> `goodpractice::gp()` yielded the following:
>
> ```
>   ✖ write unit tests for all functions, and all package code in general. 94% of code lines are
>     covered by test cases.
>
>     R/backup_object.R:73:NA
>     R/backup_object.R:75:NA
>     R/backup_object.R:76:NA
>     R/backup_object.R:81:NA
>     R/backup_object.R:82:NA
>     ... and 56 more lines
>
>   ✖ fix this R CMD check NOTE: Note: found 137 marked UTF-8 strings
> ```
>
> #### Documentation
>
> The package contains a vignette and a README on the repository, and both provide a high level overview how one would use it. I think most potential users could get started right away. Every exported function is documented and has examples of how to use them.
>
> I think a more extensive vignette with examples for more of the exported functions would be helpful. `taxlist` has a lot of exported functions, and the bulk of them are well documented and demonstrated. However, some have few details on how one might use some of them in an actual workflow. For example, the `replace_*`, `insert_rows`, and `update_*` functions aren't referenced at all in the vignette or Readme. The examples in the documentation are helpful to some degree, but for the first two, don't necessarily demonstrate how one would use them in a `taxlist` workflow.
>
> A couple small notes:
>
>     1. for functions with more than 1 or 2 arguments, I would also suggest naming all of the function arguments in the examples section. It's a small thing, but scrolling up and down between the `Usage` and `Examples` section to figure out what each argument corresponds to in the examples can be frustrating. It appears that this is done for the functions that are most frequently used (and referenced in the vignette), but not for others (e.g. `replace_*`, `insert_rows`).
>
>     2. I'd suggest moving the code to access the vignette from the vignette itself to the Readme and/or the startup message (but check the rOpenSci documentation guidelines first, the latter may not best practice actually).
>
>
> I would expect the vignette enhancements could take some time, so I view that as more of a long term issue than something that _must_ be corrected for this review.
> #### Functions
>
> These generally work as I would expect them to. My primary critique of the code is the usage of `with()` in the source code for the functions and methods. This is generally recommended against in programming contexts (see that function's documentation and section 6 of [Thomas Lumley's Guide to NSE](https://developer.r-project.org/nonstandard-eval.pdf)). In this package, it introduces an unexpected behavior in the `$` method, which I explain below. I was not able to generate other instances where the `with(...)` would fail silently for the other methods it appears in, but I haven't tested this exhaustively either.
>
> I think the `$` method a cool idea for quickly accessing some slots, and could help users that aren't familiar with the S4 `@` usage to dive right in. On the other hand, it is somewhat confusing to more experienced users who won't expect to see a `$` call on an S4 object, particularly if they're reviewing code written by someone else that makes use of this package. If you decide to keep it, then I'd recommend making its behavior more consistent and having it throw an informative error that tells the user exactly which slots it can be used on. For example:
>
> ```
> >data(Easplist)
>
> # In Easplist@taxonNames
> >Easplist$TaxonName
> Error in get(name) : object 'TaxonName' not found
>
> # Isn't present in any Easplist slots, but is in NULLing.R
> >Easplist$score
> NULL
>
> # Is present in Easplist@taxonNames and NULLing.R,
> # but not taxonTraits or taxonRelations
>
> >Easplist$TaxonUsageID
> NULL
>
> # Is present in Easplist@taxonNames,
> # but not taxonTraits,  taxonRelations or NULLing.R
>
> >Easplist$AuthorName
> Error in get(name): object 'AuthorName' not found
> ```
>
> It looks like the method is finding the object `TaxonUsageID` in the `taxlist` namespace which is set in `NULLing.R`. This is because the `inherit` argument in `get(name)` defaults to `TRUE`, and it isn't altered in the method's source code. `Easplist$acceptedname` and `Easplist$score` also yield `NULL`, while `Easplist$TaxonConceptID` is found in the environment generated by `with`, and so no further searching is performed by `get`. These are the only names generated in `NULLing.R`. However, if one created a new column in one of the tables, then tried to access it, we get the classic:
>
> ```
>
> Easplist@taxonNames$mean <- runif(nrow(Easplist@taxonNames))
>
> > Easplist$mean
> object of type 'closure' is not subsettable
> ```
>
> I'd suggest adding something like this to the beginning of the `$` method source code:
>
> ```
>
> pos_nms <- c(names(x@taxonTraits),
>              names(x@taxonRelations))
>
> if(! name %in% pos_nms && name %in% c(names(other_tables))) {
>
>   stop("can only access 'taxonTraits' or 'taxonRelations' slots with '$'")
>
> }
> ```
>
>     * Side note: the documentation for the `$` indicates that `$<-` method is implemented, but I can't find the source code for it and I get the following error when I try to use: `"no method for assigning subsets of this S4 class`". Does this mean it is not yet implemented?
>
>
> Note that none of the above is a critique of using `with()` in examples/vignettes/Readme. Those are interactive cases, and so I don't see any major issues there, except that it may confuse some users who are not familiar with the function itself.
>
> Finally, it would be nice to have a more standardized function naming system across the package. For example, some are named `replace_*`, and others named `update_*`, and as best I can tell, they have the same higher level purpose (i.e. modifying an object w/ new information). The rOpenSci package guidelines suggest an `object_verb` naming scheme (e.g. something like `tl_update_concept()`, `tl_add_concept`, `tl_update_idx()`, `tl_insert_rows()`, etc.).
> ### Minor comments
>
>     * rOpenSci guidelines suggest not including "This is package_vX.Y.Z" in the startup message.
>
>     * More of a wish, but a `print.taxlist` or `show.taxlist` method would be nice, even if it's identical to the current `summary` method.
>
>     * Could the `rm.na`/`include_na` arguments in `count_taxa` switch to `na.rm` for consistency w/ other `base`/`stats` functions?
>
>     * Couldn't find contributing/community guidelines in the repository.
>
>     * I hope this isn't annoying, but I would suggest using a code format that is more consistent with the norms in the R community (e.g. [Google R style guide](https://google.github.io/styleguide/Rguide.html), or the [tidyverse style guide](https://style.tidyverse.org/index.html)). This will enable other contributors to more easily read and understand the source code, and reduce the difficulty in making contributions. Specifically, I'd suggest
>
>       * spacing after all commas, and after opening `[` when the first argument is empty (i.e. subsetting columns).
>       * single statements on a single line except where it is impossible to keep them under 80 characters wide.
>       * indenting that matches opening brackets/parentheses in function calls.
>
>     * Description lists the following packages in Suggests, but I don't see them used in the vignettes, Readme, or Code:
>
>       * `stringi`
>       * `goodpractice`
>       * It looks like these were included in the `data-raw/0_check.R` file. I don't think they need to be included in the package though, as they are already `.Rbuildignore`'d since they're in `data-raw/` and so shouldn't cause any problems with `check` and/or CRAN. I had no issues when re-running `check` with them removed.
>
>     * Consider making a package website. My experience w/ using [`pkgdown`](https://pkgdown.r-lib.org/index.html) for this purpose has been excellent.
>
>
> Excellent work @kamapu and @zachary-foster! Let me know if I can help with any of the points above, or if I missed any specific points in the review guide!

# Response to mcsiple

> Hi @maelle, @kamapu, and @zachary-foster, here's my review! Please let me know if you have any questions about anything I wrote.
> ## Package Review
>
> _Please check off boxes as applicable, and elaborate in comments below. Your review is not limited to these topics, as described in the reviewer guide_
>
>     * **Briefly describe any working relationship you have (had) with the package authors.**
>
>     * [x]  As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).
>
>
> #### Documentation
>
> The package includes all the following forms of documentation:
>
>     * [ ]  **A statement of need** clearly stating problems the software is designed to solve and its target audience in README
>
>     * [x]  **Installation instructions:** for the development version of package and any non-standard dependencies in README
>
>     * [x]  **Vignette(s)** demonstrating major functionality that runs successfully locally
>
>     * [x]  **Function Documentation:** for all exported functions in R help
>
>     * [x]  **Examples** for all exported functions in R Help that run successfully locally
>
>     * [ ]  **Community guidelines** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`).
>       I did not find any Contributing Guidelines in the README or DESCRIPTION
>
>
> > ##### For packages co-submitting to JOSS
> >
> > * [ ]  The package has an **obvious research application** according to [JOSS's definition](https://joss.theoj.org/about#submission_requirements)
> >
> > The package contains a `paper.md` matching [JOSS's requirements](https://joss.theoj.org/about#paper_structure) with:
> >
> > * [ ]  **A short summary** describing the high-level functionality of the software
> > * [ ]  **Authors:**  A list of authors with their affiliations
> > * [ ]  **A statement of need** clearly stating problems the software is designed to solve and its target audience.
> > * [ ]  **References:** with DOIs for all those that have one (e.g. papers, datasets, software).
>
> #### Functionality
>
>     * [x]  **Installation:** Installation succeeds as documented.
>
>
> ###### Local install
>
> Local install worked fine; I had some classic issues with updating {rlang} and {ps} and a couple of other packages but once I had all the dependencies there weren't any problems installing locally. (I say "classic" issues because I have often had problems when {rlang} needs to be updated. Not sure why this is.)
> ###### Install from GitHub
>
> Install from GitHub works smoothly.
>
>     * [x]  **Functionality:** Any functional claims of the software been confirmed.
>
>
> ###### Good practices checklist
>
> `r goodpractice::gp()` returns three good practices that are missing from {taxlist}:
>
>     * fix this R CMD check NOTE: Note: found 137 marked UTF-8 strings
>
>     * fix this R CMD check WARNING: LaTeX errors when creating PDF version. This typically indicates Rd problems.
>
>     * fix this R CMD check ERROR: Re-running with no redirection of stdout/stderr. Hmm ... looks like a package Error in texi2dvi(file = file, pdf = TRUE, clean = clean, quiet = quiet, : pdflatex is not available Error in texi2dvi(file = file, pdf = TRUE, clean = clean, quiet = quiet, : pdflatex is not available Error in running `tools::texi2pdf()`
>
>
> I am not sure what exactly is happening with any of these three but it sounds like an issue with PDF creation, which has come up before (see "Checks on package source").
> ###### Metadata files
>
>     * #### [README](https://github.com/kamapu/taxlist)
>
>     * #### [DESCRIPTION](https://github.com/kamapu/taxlist/blob/master/DESCRIPTION)
>
>     * #### [NAMESPACE](https://github.com/kamapu/taxlist/blob/master/NAMESPACE)
>
>     * [x]  **Performance:** Any performance claims of the software been confirmed.
>
>     * [ ]  **Automated tests:** Unit tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.
>       Some of the tests failed when I ran test(pkg_dir) (See "Package tests & checks" section)
>
>     * [x]  **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines
>
>
> #### Final approval (post-review)
>
>     * [ ]  **The author has responded to my review and made changes to my satisfaction. I recommend approving this package.**
>
>
> Estimated hours spent reviewing: 6
>
>     * [x]  Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer ("rev" role) in the package DESCRIPTION file.
>
>
> #### Key review checks
>
>     * [x]   Does the code comply with **general principles in the [Mozilla reviewing guide](https://mozillascience.github.io/codeReview/review.html)**?
>
>     * [x]   Does the package **comply with the [ROpenSci packaging guide](https://ropensci.github.io/dev_guide/building.html)**?
>
>     * [x]   Are there **improvements** that could be made to the **code style?**
>
>     * [ ]   Is there **code duplication** in the package that should be reduced?
>
>     * [ ]   Are there **user interface improvements** that could be made?
>
>     * [ ]   Are there **performance improvements** that could be made?
>
>     * [x]   Is the [**documentation**](https://ropensci.github.io/dev_guide/building.html#documentation) (installation instructions/vignettes/examples/demos) **clear and sufficient**?
>       (Note: the )
>
>
> ### Review Comments
>
> Thank you for the opportunity to review this package. It seems to be used very widely already, so I know it is a useful tool. Like @levisc8 , I don't do a lot of taxonomy. I tend to work with existing taxonomic classifications for things (and use taxonomy more as a unique ID to get other information like stock status), but I tried to review the package as thoroughly and usefully as I could, but I urge the package authors to interpret my comments with that lens, and not to struggle to incorporate certain changes if you know they won't improve the functionality for everyday users of the package, or if it seems that I have missed the point on that topic.
> #### Package checks & tests
> ##### Checks
>
> 0 errors and 1 warning returned from `r devtools::check()` : "`The warning is "'qpdf' is needed for checks on size reduction of PDFs`" (probably not important, but just so you have it!)
> ##### Tests
>
> I ran into a couple errors and warnings in this section, which I see @levisc8 has suggested some useful solutions for. I think they're just coming from how the directories are set up for the checks and the examples.
>
> `devtools::check(pkg_dir)` returns 4 failures and 2 warnings.
>
> I got 4 failures:
>
>     * R:3 coerce objects to taxlist
>
>     * R:4 loading a backup ("requested backup missing from working directory")
>       cannot open compressed file for examples.Rda
>       It looks like examples.Rda is not where check() is looking for it.
>       Note: I think this may be because the directory in test-taxon_relations.R should go to inst/taxlist_examples/examples.Rda. Currently that line is missing the inst/ part.
>
>     * R:5 unable to open DBF file in importing Turboveg datasets
>       I'm not sure why this is happening. It looks like it's set up properly in the test (no directory issues). I saw one little post online that said sometimes Windows expects the .DBF to be capitalized in the file extension but changing it manually did not affect the test outcome for me. If I had more time I would test on a Mac too, and get to the bottom of this!
>
>
> I got 3 warnings from `devtools::check()`:
>
>     * R:3 "no such file as cyperus/names.csv"
[…]
