### 2020-07-02 #233 @kamapu
@levisc8  I have some comments regarding your review, which I would like to clarify before editing my response:

1. Regarding `devtools::check()`. Everything is running OK in my case. May this problem be related with the fact that data have a different relative path in the source as in the installed package? In any case, if your offer of solving it in the source is still open, then go ahead.

2. In `goodpractice::gp()`, the message *fix this R CMD check NOTE: Note: found 137 marked UTF-8 strings*. I don't really now, how to solve this issue or if it is an issue at all. I already had solved similar problems reported by CRAN in https://github.com/kamapu/taxlist/commit/3b2269ddb05ec01baebb7afec566b2c1d47a8665

3. In the case of `data-raw/0_check.R`, this is the script I use to automatically build and check `taxlist` before committing changes to GitHub (honestly, I'm sort of proud about it). I know, it may be confusing for contributors but I have to keep it in the source. Is there a better place and/or a better name for this script?
