+++ DESCRIPTION
@@ -6,7 +6,9 @@ Authors@R: c(
     person("Miguel", "Alvarez", email="kamapu78@gmail.com",
         role=c("aut","cre"), comment=c(ORCID="0000-0003-1500-1834")),
     person("Zachary", "Foster", email="zacharyfoster1989@gmail.com",
-        role="ctb", comment=c(ORCID = "0000-0002-5075-0948"))
+        role="ctb", comment=c(ORCID = "0000-0002-5075-0948")),
+    person("Sam", "Levin", email="levisc8@gmail.com", role="rev"),
+	person("Margaret", "Siple", email="siplem@uw.edu", role="rev")
     )
 Depends:
     R(>= 3.5.0)
