### 2020-07-02 #233 @kamapu
I forgot to mention it, @levisc8 : I work with **eclipse + StatET** (not RStudio), thus I need the `data-raw/0_check.R` to carry out builds and tests.
