### 2020-07-03 #233 @kamapu
Dear @levisc8
In your review, under **Functionality**, there are two unchecked boxes but without comments:

> * [ ] **Automated tests:** Unit tests cover essential functions of the package
> * [ ] **Packaging guidelines:** The package conforms to the rOpenSci packaging guidelines.

I'm wondering if you may have forgotten to check them (in that case, I just ignore them) or if there are issues to be solved on that regard.
