### 2020-07-03 #233 @kamapu
... this questions is for all reviewers + editor: @levisc8 @mcsiple @maelle

`goodpractice::gp()` will complain that:

1. The coverage is not 100%.
2. Issues with UTF-8 strings, although the encoding is declared in the description and there is no NOTE on this regard when using R CMD check in a Latin-1 environment, as required by CRAN.

Am I allowed to ignore these issues in the review? I mean, I can redact a brief response but no changes in code will be required.
