### 2020-07-03 #233 @kamapu
@mcsiple : In your outputs of `goodpractice::gp()`, the LaTeX error can be related to your local LaTeX distribution or your installed version of `rmarkdown`.
Perhaps it can be solved by updating your LaTeX distribution and/or `rmarkdown`. The later from the GitHub repository???
