### 2020-07-03 #233 @levisc8
Hi @kamapu ,
Thanks for getting to this so quickly! I'll respond in order:

1. Re `devtools::check()`: I'll try to get a pull request together by the middle of next week. The issue I had was not with `check` itself, but running `devtools::test` from a command prompt. The tests do not work interactively, and the line highlighted in @maelle 's comment [here](https://github.com/ropensci/software-review/issues/233#issuecomment-643245674) is exactly the fix I was suggesting. I believe that altering the tests that use that (with the exception of the one I mentioned in my review) will resolve that issue.

    - I'm not sure how huge of an issue this is since I don't think very many people run `devtools::test()` on other people's source packages frequently (if ever). I did think it was worth pointing out though, as the tests do not technically pass on my local machine (which is why I didn't check the box in my review).

2.  It appears to come from some of the example data sets - you could try something like [this](https://stackoverflow.com/questions/29043932/how-to-handle-example-data-in-r-package-that-has-utf-8-marked-strings) to set it right. I *think* that `check` gets fussy if there are *any* non-ASCII strings in a package, but my understanding of locales and encodings is very limited, so I don't know why it would get fussy about this.

    - As for the test coverage - I don't know what rOpenSci's requirements are, but I don't expect 100% coverage. 94% is pretty darn good :)

3. It is a cool script! I don't mean to say "get rid of the script". You don't need to move it either, as the `data-raw` directory is in your `.rbuildignore`, and so will never find its way into a built package. I just meant that I don't think you need those packages in the `Suggests` field of your DESCRIPTION file. The reason for this is that some users may want to install all dependencies, but not want to install unneeded packages for `taxlist`'s functionality (particularly stringi and devtools, which can take some time to install from source). Sorry for the confusion! Does that make sense?

    - I have no problems w/ checking if I remove devtools. stringi, and goodpractice from the Suggests field, so I don't think it will cause any issues w/ CRAN or other OS's. It just makes taxlist a bit leaner is all.

4. Packaging guidelines - I think I left that un-checked because of the website, but that is clarified now so I've marked it complete :)
