### 2020-07-03 #233 @kamapu
@levisc8 I was just reading your comments again:
>     1. Regarding `devtools::check()`. Everything is running OK in my case. May this problem be related with the fact that data have a different relative path in the source as in the installed package? In any case, if your offer of solving it in the source is still open, then go ahead.

I guess, I'm able to do the proposed change and test it.
