+++ DESCRIPTION
@@ -22,14 +22,10 @@ Imports:
     utils,
     vegdata
 Suggests:
-    devtools,
-    goodpractice,
     knitr,
-    stringi,
     testthat,
     rmarkdown,
-    roxygen2,
-    covr
+    roxygen2
 LazyData: true
 Description: Handling taxonomic lists through objects of class 'taxlist'.
     This package provides functions to import species lists from 'Turboveg'
