### 2020-07-03 #233 @kamapu
@levisc8
Just a comment on your comment:
> 2. It appears to come from some of the example data sets - you could try something like this to set it right...

The encoding issues have been the biggest nightmare not only when programming but also in real life (see [here](https://github.com/ropensci/RefManageR/issues/61#issuecomment-478556229)).
I pretend to assume, that the issue was already solved before and `goodpractice::gp()` is missing some settings to tolerate UTF-8 encodings.
