### 2020-07-03 #233 @kamapu
@levisc8
> you can replace the file.path(path.package(...)) with system.file('dir_within_pkg', 'file_name', package = 'taxlist')

This solves the problems in `devtools::test()` but causes errors in `devtools::check_built()`
