### 2020-07-04 a8b2c6a
StartMessage deleted
