+++ tests/testthat/test-df2taxlist.R
@@ -1,7 +1,9 @@
 context("coerce objects to taxlist")
 
 Cyperus <- read.csv(file.path(path.package("taxlist"), "cyperus", "names.csv"),
-		stringsAsFactors=FALSE)
+        stringsAsFactors=FALSE)
+
+system.file('dir_within_pkg', 'file_name', package = 'taxlist')
 
 test_that("coercion to taxlist works", {
 			expect_is(df2taxlist(Cyperus, AcceptedName=!Cyperus$SYNONYM),
