+++ data-raw/0_check.R
@@ -8,6 +8,7 @@ library(devtools)
 library(covr)
 library(goodpractice)
 library(rmarkdown)
+library(knitr)
 
 # Document package
 document()
@@ -15,11 +16,11 @@ document()
 # Report coverage
 report()
 
-# Render readme-file.
-render("README.Rmd")
-
 # Write data set
 source("data-raw/Easplist/Easplist.R")
+
+# Purl vignette R-code
+purl("vignettes/taxlist-intro.Rmd", "vignettes/taxlist-intro.R")
 
 # Check application of good practices
 gp()
@@ -31,3 +32,12 @@ Ploc <- build(path=file.path(Root, "00_Rpackages"))
 # Test the package
 Sys.setenv(LANG="en_US.iso88591")
 check_built(path=Ploc)
+
+# After check ------------------------------------------------------------------
+
+# TODO: Additional steps at final version
+# Install the package
+# Build vignette for homepage
+
+# Render readme-file.
+render("README.Rmd")
