### 2020-07-04 af5ed50
New methods show() and print() for taxlist objects
