+++ R/summary.R
@@ -132,13 +132,15 @@ overview_taxon <- function(object, ConceptID, display, maxsum, secundum=NULL) {
 
 #' @name summary
 #' 
+#' @rdname summary
+#' 
 #' @title Print overviews for taxlist Objects and their content
 #' 
 #' @description 
 #' A method to display either an overview of the content of
 #' [taxlist-class] objects or an overview of selected taxa.
 #' 
-#' @param object A [taxlist-class] object.
+#' @param object,x A [taxlist-class] object.
 #' @param ConceptID IDs of concepts to be displayed in the summary.
 #' @param units Character value indicating the units shown in the object's
 #'     allocated space.
@@ -198,4 +200,38 @@ setMethod("summary", signature(object="taxlist"),
                 overview_taxlist(object, units, check_validity) else
                 overview_taxon(object, ConceptID, display, maxsum, secundum)
         }
+)
+
+#' @rdname summary
+#' 
+#' @aliases show show,taxlist-method
+#' 
+#' @exportMethod show
+#' 
+setMethod("show", signature=(object="taxlist"),
+		function(object) {
+			summary(object)
+		}
+)
+
+#' @rdname summary
+#' 
+#' @aliases print
+#' 
+#' @exportMethod print
+#' 
+if(!isGeneric("print"))
+	setGeneric("print",
+			function(x, ...)
+				standardGeneric("print")
+	)
+
+#' @rdname summary
+#' 
+#' @aliases print,taxlist-method
+#' 
+setMethod("print", signature=(x="taxlist"),
+		function(x, ...) {
+			summary(x, ...)
+		}
 )
