### 2020-07-05 d35e7e0
Improvements in '$' method
