+++ R/Extract.R
@@ -7,19 +7,16 @@
 #' [taxlist-class] objects.
 #' 
 #' @param x Object of class [taxlist-class].
-#' @param name A name to access.
-#' @param i,j Indices for access.
+#' @param i Integer or logical vector used as index for access to taxon
+#'   concepts, referring to the rows in slot 'taxonRelations'. These indices can
+#'   be used to produce a object with a subset of taxon concepts. It is not
+#'   recommended to use character values for this index.
+#' @param j Integer, logical or character vector used as index for access to
+#'   variables in slot 'taxonTraits'. These indices can be used to reduce the
+#'   number of variables in the mentioned slot.
 #' @param drop A logical value passed to \code{\link[base]{Extract}}.
-#' 
-#' @details 
-#' While the method `$` automatically recognizes the slot queried, provided
-#' that there is no shared column names.
-#' 
-#' In the method `[`, the first index is referred to the rows in slot
-#' `taxonRelations`, while the second index indicate the columns in slot
-#' `taxonTraits`.
-#' 
-#' A replacement method `$<-` is also implemented.
+#' @param name A symbol or character value for the method `$`, corresponding to
+#'   a variable either at slot 'taxonTraits' or slot 'taxonRelations'.
 #' 
 #' @return The method `$` retrieves a vector, while `[` retrieves a subset
 #' of the input [taxlist-class] object.
@@ -65,6 +62,10 @@ setMethod("[", signature(x="taxlist"),
 #' @exportMethod $
 #' 
 setMethod("$", signature(x="taxlist"), function(x, name) {
+			pos_nms <- c(colnames(x@taxonTraits), colnames(x@taxonRelations))
+			if(!name %in% pos_nms)
+				stop(paste("Only variables in slots 'taxonTraits' or",
+								"'taxonRelations' can be accessed by '$'"))
 			if(!paste(substitute(name)) %in% colnames(x@taxonRelations))
 				x@taxonRelations[[name]] <- with(x@taxonTraits,
 						get(name)[match(x@taxonRelations$TaxonConceptID,
