### 2020-07-05 8eb93f7
fix file paths in tests
