+++ R/backup_object.R
@@ -46,8 +46,8 @@
 #' @examples 
 #' \dontrun{
 #' ## A subset with Pseudognaphalium and relatives
-#' Pseudognaphalium <- subset(Easplist, grepl("Pseudognaphalium", TaxonName),
-#' slot="names")
+#' Pseudognaphalium <- subset(x=Easplist, subset=grepl("Pseudognaphalium",
+#'     TaxonName), slot="names")
 #' Pseudognaphalium <- get_parents(Easplist, Pseudognaphalium)
 #' 
 #' ## Create a backup with date stamp
