+++ R/match_names.R
@@ -43,7 +43,7 @@
 #' match_names(species, Easplist)
 #' 
 #' ## Display accepted names in output
-#' match_names(species, Easplist, show_concepts=TRUE)
+#' match_names(x=species, object=Easplist, show_concepts=TRUE)
 #' 
 #' @rdname match_names
 #' 
