+++ R/print_name.R
@@ -37,22 +37,22 @@
 #' summary(Easplist, 363, secundum="secundum")
 #' 
 #' ## Empty plot
-#' plot(NA, xlim=c(0,5), ylim=c(7,1), bty="n", xaxt="n", xlab="",
+#' plot(x=NA, xlim=c(0,5), ylim=c(7,1), bty="n", xaxt="n", xlab="",
 #'     ylab="options")
 #' 
 #' ## Accepted name with author
-#' text(0, 1, labels=print_name(Easplist, 363, style="expression"), pos=4)
+#' text(x=0, y=1, labels=print_name(Easplist, 363, style="expression"), pos=4)
 #' 
 #' ## Including taxon view
-#' text(0, 2, labels=print_name(Easplist, 363, style="expression",
+#' text(x=0, y=2, labels=print_name(Easplist, 363, style="expression",
 #'     secundum="secundum"), pos=4)
 #' 
 #' ## Second mention in text
-#' text(0, 3, labels=print_name(Easplist, 363, style="expression",
+#' text(x=0, y=3, labels=print_name(Easplist, 363, style="expression",
 #'     second_mention=TRUE), pos=4)
 #' 
 #' ## Using synonym
-#' text(0, 4, labels=print_name(Easplist, 50037, style="expression",
+#' text(x=0, y=4, labels=print_name(Easplist, 50037, style="expression",
 #'     concept=FALSE), pos=4)
 #' 
 #' ## Markdown style
@@ -62,7 +62,7 @@
 #' text(0, 6, labels=print_name(Easplist, 363, style="html"), pos=4)
 #' 
 #' ## LaTeX style for knitr
-#' text(0, 7, labels=print_name(Easplist, 363, style="knitr"), pos=4)
+#' text(x=0, y=7, labels=print_name(Easplist, 363, style="knitr"), pos=4)
 #' 
 #' @rdname print_name
 #' 
