+++ R/accepted_name.R
@@ -56,8 +56,8 @@
 #' 
 #' ## Inserting a new name first
 #' summary(Easplist, "Basella alba")
-#' Easplist <- add_synonym(Easplist, 68, TaxonName="Basella cordifolia",
-#'     AuthorName="Lam.")
+#' Easplist <- add_synonym(taxlist=Easplist, ConceptID=68,
+#'     TaxonName="Basella cordifolia", AuthorName="Lam.")
 #' summary(Easplist, 68)
 #' accepted_name(Easplist, 68) <- 56139
 #' summary(Easplist, 68)
