### 2020-07-05 a6396a4
Improvements in the text of vignette
