### 2020-07-06 #233 @maelle
Hello @kamapu

* Regarding test coverage we don't expect more than 75% coverage https://devguide.ropensci.org/building.html#testing

* If the encoding issue is a problem with goodpractice, it might be worth opening an issue in goodpractice repo itself? You could also run a check on e.g. an R-hub platform with a different encoding setting https://blog.r-hub.io/2019/04/25/r-devel-linux-x86-64-debian-clang/#r-devel-linux-x86_64-debian-clang just to make sure you don't get any check NOTE/WARNING there.

* Just curious why you use Eclipse, historical reasons or some advantages? No answer required, this is a bit off-topic.
