### 2020-07-06 295ed95
Purged use of 'with()' in code and vignette
