### 2020-07-06 807fe5e
Empty spaces between square-brackets when an index is omitted
