+++ tests/testthat/test-taxon_traits.R
@@ -4,7 +4,7 @@ test_that("function taxon_traits is working", {
 			expect_is(taxon_traits(Easplist), "data.frame")
 			expect_error(taxon_traits(Easplist) <- iris)
 			tmp <- Easplist
-			tmp_traits <- Easplist@taxonTraits[1:10,]
+			tmp_traits <- Easplist@taxonTraits[1:10, ]
 			expect_equal(nrow({taxon_traits(tmp) <- tmp_traits
 							taxon_traits(tmp)}) == nrow(tmp_traits), TRUE)
 			tax_1 <- df2taxlist(data.frame(TaxonName="Poa annua",
