+++ R/taxon_views.R
@@ -119,7 +119,7 @@ setMethod("add_view", signature(taxlist="taxlist"),
 				old_view <- taxlist@taxonViews
 				for(i in colnames(new_view)[!colnames(new_view) %in%
 								colnames(old_view)]) {
-					old_view[,i] <- rep(NA, nrow(old_view))
+					old_view[ ,i] <- rep(NA, nrow(old_view))
 				}
 				new_view <- do.call(rbind, list(old_view, new_view))
 			}
