### 2020-07-07 #233 @kamapu
Hello @maelle

1. All right.

2. Unfortunately I got the same output at r-hub:

> * checking data for non-ASCII characters ... NOTE
  Note: found 137 marked UTF-8 strings

3. A bit of both: I'm been working for long time with eclipse. The biggest advantages over RStudio are as well the representation of objects in "Object browser" and the possibility of work with a broad spectrum of applications (R, Web Development, LaTeX, PostgreSQL, etc.) in just one workbench. Otherwise I use RStudio to introduce R in our courses.
