### 2020-07-07 #233 @maelle
It's unfortunate but means the problem needs to be fixed (or well justified), because you'd get it on CRAN check platform that has this encoding setting. Would you mind opening a question about that at https://discuss.ropensci.org/?
