### 2020-07-07 #233 @kamapu
Hello all: @maelle @levisc8 @mcsiple

I see, I have to dedicate some time with the example data, again. I though, I have solved it in the last release at CRAN but I just realized, the test is still providing the warning about UTF-8 Strings.

I assume, **R CMD check** is not giving any warning because of my local settings (I'm working in Linux).

Perhaps some of you could give me some hints:

1. The data is originally stored in PostgreSQL, also using UTF-8 encoding. I export the data as csv-file in `data-raw/Easplist`, also using UTF-8. And then I create the `taxlist` object in an R-Session. Perhaps something can be adjusted in this workflow to get rid of the warnings.

2. I was assuming that using `Sys.setenv(LANG="en_US.iso88591")` I was testing the package in the CRAN settings but it is evidently not working.
