### 2020-07-07 #233 @kamapu
I always advice our students to not confuse ERRORS with WARNINGS. And now I see me confusing WARNINGS with NOTES. Are the later usually tolerated?
