### 2020-07-07 #233 @maelle
they could but they need to be justified, and their mere existence delays the CRAN review process, that's why it's better to get rid of them.
