### 2020-07-07 #233 @kamapu
@maelle I see your point, although I don't really know why UTF-8 is a problem. Anyway, the issue is that the vector including author names is prone to have "special characters", actually even taxon usage names (for instance the genus *Isöetes* and its species). In `taxlist` the combination between taxon usage names and authority (columns **TaxonName** and **AuthorName** in slot **taxonNames**) is a sort of identity for a name (no duplicated allowed).

I'm willing to solve the problem but I will need support on this regard, especially because I wish, the package can still be of any use disregarding the local settings of the users. I also wish to reach THE solution, because the issue is emerging again, and again, ... :sob: :sob: :sob:
