### 2020-07-07 #233 @maelle
To clarify, I am not saying it's a bad NOTE but I think it'd be easier if you opened a thread at https://discuss.ropensci.org where more community members could chime in with ideas for making the NOTE disappear or writing a good comment for CRAN (who'd be the ones to be convinced the NOTE is fine), from their own experience.

I do understand encoding pain, my name is Ma?#$lle I mean Maëlle. :wink:
