### 2020-07-07 #233 @kamapu
@maelle Thanks for clarification. I sort of lost the path (cause I'm doing many things in parallel). I'll do as suggested but not right now...

Perhaps it was a premonition but long time ago I decided to skip the accent from Álvarez when self-referencing...
