### 2020-07-08 f7341af
Improved documentation for 'replace_*()' and 'insert_rows()'
