+++ R/replace_x.R
@@ -3,48 +3,65 @@
 #' @title Data manipulation.
 #' 
 #' @description
-#' Functions provided for fast replacement and update of data.
+#' This is a series of functions designed for a fast coding of replacements
+#' both, as internal functions and in workflows dealing with information stored
+#' in vectors and data frames.
+#' Such functions are especially useful when handling with functional traits
+#' stored in [taxlist-class] objects.
 #' 
-#' @param x A vector to be modified or a data frame in the case of
-#'     `insert_rows`.
-#' @param old,new Vectors containing the values to be replaced and the updated
-#'     values, respectively.
-#' @param idx1,idx2 Indices applied for the values in 'x' and the values to be
-#'     replaced, respectively.
-#' @param y Data frame including rows to be inserted in `x`.
+#' `replace_x()` is used to exchange values in vectors.
+#' `replace_idx()` changes values in vectors by matching indices or conditions.
+#' The function `replace_na()` works in the same way as `replace_idx()` but will
+#' only insert values in empty elements (NAs).
 #' 
-#' @details 
-#' These are functions implemented for efficient coding of insert and replace
-#' routines.
+#' The function `insert_rows()` will add rows and columns at the same time.
+#' This function will be used when a new table is appended to another but
+#' sharing only part of the columns.
 #' 
-#' The functions `replace_x` and `replace_idx` replace values in vectors,
-#' in the first case comparing values in the vector and in the second one by
-#' using indices.
-#' The function `replace_na` works in the same way as `replace_idx`, but
-#' carries out the replacement only if the previous value is a `NA`.
-#' 
-#' The function `insert_rows` inserts `y` as new rows in `x`. If `y` contains
-#' columns absent in `x`, they will be added to the output data frame.
+#' @param x A vector to be modified. In the case of `insert_rows()`, `x` is a
+#'     data frame.
+#' @param old A vector with values to be replaced by `replace_x()` in a vector.
+#' @param new A vector containing values to be inserted, either comparing values
+#'     or using indices.
+#' @param idx1,idx2 Indices applied for value replacements to match `x` with
+#'     `new`, respectively.
+#' @param y A data frame including rows (and columns) to be inserted in `x`.
 #' 
 #' @return A vector or data frame with the modified values.
 #' 
 #' @author Miguel Alvarez.
 #' 
 #' @examples 
-#' library(taxlist)
-#' 
 #' ## Replace values in vector
 #' replace_x(x=letters, old=c("b", "p", "f"), new=c("bee", "pork", "fungus"))
 #' 
 #' ## Replace values using indices
 #' replace_idx(x=letters, idx1=1:length(letters), idx2=c(2, 7, 17),
-#'     new=c("second", "seven", "seventeenth"))
+#'     new=c("second", "seventh", "seventeenth"))
 #' 
 #' ## Replace values if they are NAs
 #' letters[2] <- NA
 #' replace_na(x=letters, idx1=1:length(letters), idx2=c(1:3),
 #'     new=c("alpha", "beta", "zeta"))
 #' 
+#' ## The same applications but this time for functional traits
+#' summary(as.factor(Easplist$lf_behn_2018))
+#' 
+#' # Merge annuals
+#' Easplist@taxonTraits$lifeform <- replace_x(
+#'     x=Easplist@taxonTraits$lf_behn_2018,
[…]
