+++ pkgdown.yml
@@ -2,5 +2,5 @@ pandoc: 2.7.3
 pkgdown: 1.5.1.9000
 pkgdown_sha: caf7cc74e008e8d0d6e53eeda93cd3cad3545bf2
 articles: []
-last_built: 2020-07-07T21:20Z
+last_built: 2020-07-08T08:41Z
 
