### 2020-07-08 e6e8c8d
Improved documentation for 'update_*()'
