+++ R/taxon_names.R
@@ -37,7 +37,7 @@
 #' ## Display of slot 'taxonNames'
 #' Euclea <- subset(x=Easplist, subset=charmatch("Euclea", TaxonName),
 #'     slot="names", keep_children=TRUE)
-#' summary(Euclea)
+#' Euclea
 #' taxon_names(Euclea)
 #' 
 #' ## Insert a synonym to Diospyros scabra
@@ -49,6 +49,10 @@
 #' ## Delete a synonym of Launaea cornuta
 #' summary(Easplist, "Launaea cornuta")
 #' Easplist <- delete_name(Easplist, 53821)
+#' summary(Easplist, "Launaea cornuta")
+#' 
+#' ## Hypothetical correction in author name in Launaea cornuta
+#' Easplist <- update_name(taxlist=Easplist, UsageID=355, AuthorName="L.")
 #' summary(Easplist, "Launaea cornuta")
 #' 
 #' @rdname taxon_names
