+++ man/taxon_traits.Rd
@@ -50,6 +50,15 @@ constructing \linkS4class{taxlist} objects from an empty one.
 \examples{
 head(taxon_traits(Easplist))
 
+## Updating traits for Launaea cornuta
+summary(Easplist, "Launaea cornuta")
+accepted_name(taxlist=Easplist, ConceptID=355, show_traits=TRUE)
+
+# Update
+Easplist <- update_trait(taxlist=Easplist, ConceptID=355,
+    lf_behn_2018="annual")
+accepted_name(taxlist=Easplist, ConceptID=355, show_traits=TRUE)
+
 }
 \seealso{
 \linkS4class{taxlist}.
