+++ man/taxon_names.Rd
@@ -71,7 +71,7 @@ For adding new taxon concepts as well you should use \code{\link[=add_concept]{a
 ## Display of slot 'taxonNames'
 Euclea <- subset(x=Easplist, subset=charmatch("Euclea", TaxonName),
     slot="names", keep_children=TRUE)
-summary(Euclea)
+Euclea
 taxon_names(Euclea)
 
 ## Insert a synonym to Diospyros scabra
@@ -83,6 +83,10 @@ summary(Easplist, "Diospyros scabra")
 ## Delete a synonym of Launaea cornuta
 summary(Easplist, "Launaea cornuta")
 Easplist <- delete_name(Easplist, 53821)
+summary(Easplist, "Launaea cornuta")
+
+## Hypothetical correction in author name in Launaea cornuta
+Easplist <- update_name(taxlist=Easplist, UsageID=355, AuthorName="L.")
 summary(Easplist, "Launaea cornuta")
 
 }
