### 2020-07-08 70c9109
Improvements in 'df2taxlist()'
