+++ R/df2taxlist.R
@@ -9,6 +9,8 @@
 #' @param x A data frame or a character vector with taxon names.
 #' @param AcceptedName A logical vector indicating accepted names with value
 #'     `TRUE`.
+#' @param levels A vector with the names of the taxonomic ranks. This argument
+#'     is passed to [taxlist::levels()].
 #' @param ... Additional vectors to be added as columns in slot`taxonNames`.
 #' 
 #' @details 
@@ -62,7 +64,7 @@ setGeneric("df2taxlist",
 #' @aliases df2taxlist,data.frame,logical-method
 #' 
 setMethod("df2taxlist", signature(x="data.frame", AcceptedName="logical"),
-        function(x, AcceptedName, ...) {
+        function(x, AcceptedName, levels, ...) {
             # If author names missing
             if(!"AuthorName" %in% colnames(x))
                 x$AuthorName <- NA
@@ -97,11 +99,22 @@ setMethod("df2taxlist", signature(x="data.frame", AcceptedName="logical"),
             # taxonRelations
             taxonRelations <- x[AcceptedName,c("TaxonConceptID","TaxonUsageID")]
             colnames(taxonRelations)[2] <- "AcceptedName"
-            # taxonNames
+			# In the case that ranks are provided
+			if("Level" %in% colnames(x)) {
+				taxonRelations$Level <- x$Level
+				x <- x[,colnames(x) != "Level"]
+			}
+			# In the case that parents are provided
+			if("Parent" %in% colnames(x)) {
+				taxonRelations$Parent <- x$Parent
+				x <- x[,colnames(x) != "Parent"]
+			}
+			# taxonNames
             extra_cols <- list(...)
             suppressWarnings({
                         if(length(extra_cols > 0))
-                            for(i in names(extra_cols)) x[ ,i] <- extra_cols[[i]]
+                            for(i in names(extra_cols)) x[ ,i] <-
+										extra_cols[[i]]
                     }
             )
             taxlist <- new("taxlist")
@@ -111,6 +124,8 @@ setMethod("df2taxlist", signature(x="data.frame", AcceptedName="logical"),
                 if(!i %in% colnames(taxonRelations)) taxonRelations[ ,i] <- NA
             taxlist@taxonNames <- x
             taxlist@taxonRelations <- taxonRelations
+			if(!missing(levels))
+				levels(taxlist) <- levels
             return(taxlist)
         }
 )
@@ -120,7 +135,7 @@ setMethod("df2taxlist", signature(x="data.frame", AcceptedName="logical"),
 #' @aliases df2taxlist,data.frame,missing-method
 #' 
 setMethod("df2taxlist", signature(x="data.frame", AcceptedName="missing"),
-        function(x, ...) return(df2taxlist(x, TRUE))
+        function(x, ...) return(df2taxlist(x, TRUE, ...))
 )
 
 #' @rdname df2taxlist
