+++ tests/testthat/test-taxon_traits.R
@@ -9,7 +9,7 @@ test_that("function taxon_traits is working", {
 							taxon_traits(tmp)}) == nrow(tmp_traits), TRUE)
 			tax_1 <- df2taxlist(data.frame(TaxonName="Poa annua",
 							AuthorName="L.", TaxonConceptID=1, TaxonUsageID=1),
-					AcceptedName=TRUE,2)
+					AcceptedName=TRUE)
 			expect_equal({taxon_traits(tax_1) <- data.frame(TaxonConceptID="1",
 								lf="grass")
 						nrow(tax_1@taxonTraits)
