+++ data-raw/0_check.R
@@ -41,14 +41,12 @@ check_built(path=pkg_loc)
 # Install the package
 ## install()
 
-# TODO: Build vignette for homepage
-
 # Render readme-file.
 render("README.Rmd")
 
 # Render package-site
 usethis::use_pkgdown()
-pkgdown::build_site()
+pkgdown::build_site(preview=FALSE)
 
 # Copy site
 r_path <- gsub("/taxlist", "", getwd())
@@ -57,7 +55,7 @@ pkg_path <- file.path(r_path, "kamapu.github.io", "rpkg")
 file.copy("docs", pkg_path, recursive=TRUE)
 unlink("docs", recursive=TRUE)
 
+unlink(file.path(pkg_path, "taxlist"), recursive=TRUE)
 file.rename(file.path(pkg_path, "docs"), file.path(pkg_path, "taxlist"))
 
 file.copy("README-figures", file.path(pkg_path, "taxlist"), recursive=TRUE)
-
