+++ README.Rmd
@@ -106,7 +106,7 @@ We will create an empty `taxlist` object using the function `new()`:
 library(taxlist)
 
 Fern <- new("taxlist")
-summary(Fern)
+Fern
 ```
 
 Then we have to set the respective taxonomic ranks. In such case, the levels
@@ -162,6 +162,44 @@ summary(Fern)
 summary(Fern, "all")
 ```
 
+## From data frame to taxlist
+
+A more convenient way is to create an object from a data frame including both,
+the taxon concepts with their accepted names and the taxonomic ranks with
+parent-child relationships.
+In the case of the last example, the required data frame looks like this one:
+
+```{r}
+Fern_df <- data.frame(
+		TaxonConceptID=1:4,
+		TaxonUsageID=1:4,
+		TaxonName=c("Asplenium", "Asplenium obliquum",
+				"Asplenium obliquum var. sphenoides",
+				"Asplenium obliquum var. chondrophyllum"),
+		AuthorName=c("L.", "Forster", "(Kunze) Espinosa",
+				"(Bertero apud Colla) C. Christense & C. Skottsberg"),
+		Level=c("genus", "species", "variety", "variety"),
+		Parent=c(NA, 1, 2, 2),
+		stringsAsFactors=FALSE)
+Fern_df
+```
+
+This kind of tables can be written in a spreadsheet application and imported to
+your R session.
+The two first columns correspond to the IDs of the taxon concept and the
+respective accepted name. They can be custom IDs but are restricted to integers
+in `taxlist`.
+For the use of the function `df2taxlist()`, the two first columns are
+mandatory.
+Also note that the column **Parent** is pointing to the concept IDs of the
+respective parent taxon.
+To get the object, we just use the `df2taxlist()` indicating the sequence of
+taxonomic ranks in the argument `levels`.
+
+```{r}
+Fern2 <- df2taxlist(Fern_df, levels=c("variety", "species", "genus"))
+Fern2
+```
 
 ## Similar Packages
 
