### 2020-07-08 7645e28
Mention of improvements at ROpenSci
