### 2020-07-08 #233 @levisc8
Before I get to this again, just want to clarify - do the unresolved issues refer to the UTF-8 strings in the example data set? Or something else?
