### 2020-07-08 #233 @kamapu
In general I consider as unresolved:
1) UTF-8 Note by `goodpractice::pg()` and `rhub::check(platform="debian-clang-devel")`.
2) I think, @mcsiple got several error or warning messages, in part related to LaTeX, which I'm supposing to be relate with her local settings but it should still be clarified.

Of course, I can still receive and try suggestions.
