### 2020-07-08 #233 @KevCaz
Hi @kamapu,
As I was interested in your issue, I had a look at it. Looks like the UTF-8 note you got comes from your dataset (`Easplist.dra`) (see https://github.com/dankelley/oce/issues/1211)

```
R> tools:::.check_package_datasets(".")
Note: found 137 marked UTF-8 strings
```
Something that is not checked by default by `devtools::check()` (I'm guessing here) but is checked if you do the combo `R CMD build`/`R CMD check`.

Hope this could be useful.
