### 2020-07-10 #233 @kamapu
Dear all, sorry for annoying you with the issue of encodings. While I am sort of developing a pathological aversion for this aspect of programming, for people dealing with taxonomy (including myself), the use of names and their authorities is quite crucial in the management of information, thus considered in `taxlist` as a component of the identity of a taxon concept (the accepted name of the taxon) and the taxon usage id (alternative names used to refer to a taxon), and this is the matter of my headaches: global context = special characters...

The possible solutions explored until now, just to get rid of a note, are not satisfying me in a 100%.

Now I have a request for you: The next is a block of code including the outputs in my console (I'm working in Linux with UTF-8 locale). My question is: **Do you get the exact output as in my case?**

```r
library(stringi)
library(taxlist)

Names <- Easplist@taxonNames$AuthorName[c(5299, 5021, 5019)]
Names
#> [1] "Borsch, Kai Müll. & Eb.Fisch." "Ruiz & Pav."
#> [3] "Ség."

stri_enc_mark(Names)
#> [1] "UTF-8" "ASCII" "UTF-8"

iconv(Names, "utf8", "ascii")
#> [1] NA            "Ruiz & Pav." NA

stri_enc_toascii(Names)
#> [1] "Borsch, Kai M\032ll. & Eb.Fisch." "Ruiz & Pav."
#> [3] "S\032g."

stri_trans_general(Names, "latin-ascii")
#> [1] "Borsch, Kai Mull. & Eb.Fisch." "Ruiz & Pav."
#> [3] "Seg."
```
