### 2020-07-13 #233 @maelle
For reference link to Discourse thread https://discuss.ropensci.org/t/note-on-utf-8-strings-by-goodpractice-gp/2165
