### 2020-07-27 #233 @levisc8
@kamapu regarding your last question, I do get the exact same output as you do. Furthermore, `stri_trans_general(Names, "latin-ascii") %>% str_enc_mark()` returns `'ascii' 'ascii' 'ascii'`.
