### 2020-07-27 #233 @levisc8
Review round 2:

I'm generally fine with the new form of the package. I just have some minor comments at this point.

## Unresolved issues

1. UTF-8 strings: I don't have strong feelings about this, as it is pretty far beyond my area of expertise. If you can fix it with the code blurb from above, excellent!

2. The pdflatex errors could be because pdflatex is not in the PATH variable (see [here](https://stackoverflow.com/questions/14333759/diagnosing-r-package-build-error-pdflatex-not-available)). I don't think there's much you can do about this from your side though.

    - I think the qpdf warning is similar. Basically, just need to install qpdf and add it to the PATH variable to check smoothly.

## Rest of the response

1. Documentation is looking good.

2. package API: I understand the point regarding backward compatibility. I also don't think it's unreasonable to make breaking changes if they result in substantial improvements to the user experience. Incrementing by a major version number (e.g. 0.2 -> 1.0) should signal to most users that they should expect major changes. If users *really* don't want to change their code, they can install archived versions, either from Github or from CRAN. This is something I'd defer to rOpenSci editors on though, as I don't really believe it is a deal breaker or anything.

    - Re: `aggregate`/`count_taxa`: good point - I hadn't thought about that! On the other hand, there is still an `rm.na` argument for a couple methods. This could change to `na.rm` for consistency w/ the `base`/`stats` packages. Personally, I would forget that it is `rm.na` every single time I tried to use the package. Again, I appreciate the point about back-compatibility, but this doesn't need to be a breaking change right away - you can retain the old form and deprecate it for the next few minor versions as well.
