### 2020-07-29 #233 @mcsiple
Hi @kamapu , @maelle et al.,

My apologies for the delay in responding to these new revisions and response. I have itemized them below mostly for my own mental organization (they match the structure of @kamapu 's response, I can keep track of what I have checked).

**Note/caveat**: I normally run all the checks on two different operating systems to provide better coverage in my ran everything on my other computer. This time I did my entire initial review on a Windows OS and am now checking everything again on my Mac. This could be a "good" thing because I've covered more ground, but IMO it's also suboptimal because some of the changes I've seen could potentially be caused by different software/packages instead of changes to the package. I think this is mostly the case with the LaTeX issues, but wanted to include the disclaimer.

## Statement of need
This edit is great and satisfies my question(s) about the context and uses for the package.

## Community guidelines
Newly added contributing guidelines look good.

## Good practices checklist
### UTF-8 string issues
It looks like this one is still being hashed out to some extent... I hope you find a good resolution for it. It sounds from the [ropensci.org post](https://discuss.ropensci.org/t/note-on-utf-8-strings-by-goodpractice-gp/2165/3) like maybe this will remain an issue for people with a setup that doesn't have a font that renders the non-ascii characters. Maybe a note in the vignette about whether/how this would affect functionality would be good, or better yet a custom check or error message upon install... I am sure other package developers will have this issue so you're doing a service to the community by trying to figure it out!

### LaTeX issue
This one is my fault; I normally run all the checks on two different operating systems to provide better coverage in my ran everything on my other computer. No LaTeX errors in `goodpractice::gp()` this time, so consider this part resolved.

## Automated tests
`devtools::check()` : I found no errors with `R CMD check` this time.

## Help files
- vignettes: the added text to the readme is useful; thank you for adding!
- `Cross` question: the changes to the vignette explain the structure of the dataset better - thank you.
- Codes: Your explanation is good but if you could incorporate it into the help files that would be optimal for future users!

Thanks - let me know if you have questions/comments about any of the above.
