### 2020-09-16 91ce67d
Argument rm.na replaced by na.rm in function 'count_taxa()'
