+++ DESCRIPTION
@@ -72,4 +72,4 @@ Collate:
     'Easplist-data.R'
     'taxlist-package.R'
 VignetteBuilder: knitr
-RoxygenNote: 7.1.0
+RoxygenNote: 7.1.1
