### 2020-09-16 eae874a
Additional explanations added to the documentation of taxlist-class
