### 2020-09-16 #233 @kamapu
# Response to Reviews in Round 2

Dear @maelle @mcsiple @levisc8
I'm sorry that despite the progress on the review I had to stay out of programming for a while. Now I'm back and like to thank for all valuable comments again.

I will than refer to the comments of reviewers at round 2.

# Response to @levisc8

### Unresolved Issues

- [x] UTF-8 strings
Since the option is still open, I will take it. I would prefer to keep them as long as tolerated by CRAN.

- [x] pdf-LaTeX errors
It seems to be solved.

### Rest
- [x] Documentation is looking good.
:+1:

- [x] Change argument `rm.na` trough `na.rm` in function `count_taxa()`
It was in fact a mistake from my side. Resolved in https://github.com/kamapu/taxlist/commit/91ce67d32d322ec9d8a27347f83a2a4bee2fee17

# Response to @mcsiple

Two comments to this review, otherwise I agree with the rest.

- [x] UTF-8
Yes, this will be following me for a while. Since `taxlist` is meant to be a recipient of information that is otherwise stored in a database application, the election of Encoding will probably depend on local settings and the encoding of the data source. It is interesting to experience that while our task is to provide a way to harmonize disagreement on nomenclatures for biodiversity assessments, we stumble onto Encoding issues :thinking:

- [x] Taxon codes/identifiers
I added the explanations in the help document of `taxlist-class` in https://github.com/kamapu/taxlist/commit/eae874affbbbe082c8d3c2435977dfdc64967615
