+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 1.6.1.9000
 pkgdown_sha: cc9c7b1be8573f8af3453f6e5d0422d86c916701
 articles:
   taxlist-intro: taxlist-intro.html
-last_built: 2020-09-16T16:24Z
+last_built: 2020-09-16T17:11Z
 
