### 2020-09-29 #233 @maelle
Approved! Thanks @kamapu for submitting and @mcsiple @levisc8 for your reviews! :smile_cat:

To-dos:
    * replace your package docs URL with `https://docs.ropensci.org/package_name`
