### 2020-09-30 #233 @maelle
Please ping me once you've transferred the repo so that I might give you admin rights back, and feel free to ask me any question.
Note that the invitation to the ropensci GitHub organization expires after I think one week. If you try to transfer the repo too late, tell me to re-send an invite.
