+++ build-pkg/build_n_check.R
@@ -47,17 +47,17 @@ check_built(path=pkg_loc)
 render("README.Rmd")
 
 # Render package-site
-usethis::use_pkgdown()
-pkgdown::build_site(preview=FALSE)
-
-# Copy site
-r_path <- gsub("/taxlist", "", getwd())
-pkg_path <- file.path(r_path, "kamapu.github.io", "rpkg")
-
-file.copy("docs", pkg_path, recursive=TRUE)
-unlink("docs", recursive=TRUE)
-
-unlink(file.path(pkg_path, "taxlist"), recursive=TRUE)
-file.rename(file.path(pkg_path, "docs"), file.path(pkg_path, "taxlist"))
-
-file.copy("README-figures", file.path(pkg_path, "taxlist"), recursive=TRUE)
+## usethis::use_pkgdown()
+## pkgdown::build_site(preview=FALSE)
+## 
+## # Copy site
+## r_path <- gsub("/taxlist", "", getwd())
+## pkg_path <- file.path(r_path, "kamapu.github.io", "rpkg")
+## 
+## file.copy("docs", pkg_path, recursive=TRUE)
+## unlink("docs", recursive=TRUE)
+## 
+## unlink(file.path(pkg_path, "taxlist"), recursive=TRUE)
+## file.rename(file.path(pkg_path, "docs"), file.path(pkg_path, "taxlist"))
+## 
+## file.copy("README-figures", file.path(pkg_path, "taxlist"), recursive=TRUE)
