### 2020-10-02 #233 @kamapu
# Done
- [x] Transfer to ropensci repository
- [x] Fixed links
Done in https://github.com/ropensci/taxlist/commit/d962046b6f80a7cf28a2956203ed178ffc146886
- [x] Redirect site to 'docs/ropensci'
Done in https://github.com/ropensci/taxlist/commit/fab0ca5318f142c0ac60f611f388facb355136d6
- [x] Set links in badges
Done by solving the second comment
- [x] Add codemeta.json
Done in https://github.com/ropensci/taxlist/commit/89708aaffb7bc3ea4924741ff4c04b80e28b0bc5
