### 2020-10-02 #233 @kamapu
@maelle I just transferred the repository to ROpenSci and tried to respond to all bullet points accordingly.
There is just one issue regarding to the new package site and the images. How can I include the figures to be displayed in **docs/ropensci**?
