### 2020-10-03 #233 @maelle
@kamapu I gave you admin rights again. I am not sure I follow your question. Where are those images now? What docs/ropensci deploys is a pkgdown website. I see images in https://docs.ropensci.org/taxlist/, what is missing?
