### 2020-10-03 #233 @kamapu
@maelle
In the source of the package there is a folder **README-figures**, which contains a PNG-image of the `taxlist` data schema (Figure before chapter **Installing taxlist**) but is also used to store figures generated by plotting functions (barplot at the very bottom, before **Acknowledgements**).
While these figures are visible in https://github.com/ropensci/taxlist/, they are missing in https://docs.ropensci.org/taxlist/
