### 2020-10-04 #233 @maelle
Does https://pkgdown.r-lib.org/reference/build_home.html#images-and-figures help?
