### 2020-10-04 9c1e047
Improvements in 'match_names()'
