+++ README.Rmd
@@ -11,7 +11,7 @@ output:
 knitr::opts_chunk$set(
   collapse=TRUE,
   comment="#>",
-  fig.path="README-figures/"
+  fig.path="man/figures/"
 )
 ```
 
