+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 1.6.1.9000
 pkgdown_sha: e1e46f2e13c7efdb17bc4eae40bc3779c2b0b517
 articles:
   taxlist-intro: taxlist-intro.html
-last_built: 2020-10-02T22:36Z
+last_built: 2020-10-04T16:32Z
 
