### 2020-10-04 50b5c59
Optional sorting by 'match_names()', taxlist method
