### 2020-10-04 #233 @kamapu
Oh, yes!
@maelle I used the opportunity to add some improvements:

- Function `match_names()` is sorting results by similarity
- I included within the package code (folder **inst**) all data required by the vignette.
