### 2020-10-05 #233 @maelle
Awesome! Let me plug two blog posts of mine :grin:
* https://blog.r-hub.io/2020/06/03/vignettes/
* https://blog.r-hub.io/2020/05/29/distribute-data/

In any case I think I can close this issue but feel free to ask questions here / in the forum / etc.

Thanks all again!
