### 2020-10-05 #233 @kamapu
Thanks to everyone, also from my side.
@maelle interesting posts. It seems to be, that my original idea of allocating data in a data package is not that bad at all... I wonder if this may be also a good practice for their use in examples within CRAN packages, assuming that the data-package may not be a CRAN-package...
